### 2025-03-05 #7877 @editorialbot
[PRE REVIEW]: METISSE: METhod of Interpolation for Single Star Evolution
**Submitting author:** @poojanagrawal (<a href="http://orcid.org/0000-0002-1135-984X">Poojan Agrawal</a>)
**Repository:** https://github.com/TeamMETISSE/METISSE
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.0
