### 2025-06-13 50e23e7
added metisse def for cosmic and single stars
