### 2025-06-13 572308e
switch to binary search if list size >5
