### 2025-06-13 73f4a3d
removed lower limit for Mfgb
