### 2025-06-22 1e8c8ef
add target to build Metisse as a library
