### 2025-06-23 53b09f8
added verbose output
