### 2025-06-23 0981307
added verbose output
