### 2025-06-23 8a0cc38
added imf options (uniform and kroupa) for main_metisse
