### 2025-06-24 928a233
allow AMUSE to set the dir for the tracks
