+++ src/track_support.f90
@@ -37,6 +37,8 @@ module track_support
     
     character(len = strlen):: METISSE_DIR, METALLICITY_DIR, METALLICITY_DIR_HE
 
+    character(len = strlen):: amuse_metisse_dir, amuse_metallicity_dir, amuse_metallicity_dir_he
+
     ! for use when constructing EEP distance
     logical:: weight_center_rho_T_by_Xc
     real(dp):: Teff_scale = 2d0
