### 2025-06-24 ab5b882
allow AMUSE to set the dir for the tracks
