### 2025-06-26 5e6e539
fixed a bug in locating  bgb
