+++ src/evolv_metisse.f90
@@ -1,4 +1,4 @@
-subroutine evolv_metisse(mass,max_age,ierr,id)
+subroutine evolv_metisse(mass, max_age, ierr, id)
 
     ! evolve subroutine to use metisse in standlaone mode
     ! evolves one star at a time and writes output to file
@@ -7,21 +7,21 @@ subroutine evolv_metisse(mass,max_age,ierr,id)
     use sse_support, only:time_He_MS
 
     !variable declaration
-    real(dp), intent(in):: mass,max_age
+    real(dp), intent(in):: mass, max_age
     integer, intent(out):: ierr
-    integer, intent(in), optional :: id
+    integer, intent(in), optional:: id
 
-    integer :: str,lines,old_phase,idd,io
-    real(dp):: tphys,timestep,dt,dtr
-    character(len=strlen) :: output_file
-    logical :: output, t_end
-    real(dp) :: dms, M_env,dml,x
-    type(track), pointer :: t
+    integer:: str, lines, old_phase, idd, io
+    real(dp):: tphys, timestep, dt, dtr
+    character(len = strlen):: output_file
+    logical:: output, t_end
+    real(dp):: dms, M_env, dml, x
+    type(track), pointer:: t
 
-    ! dummy variables for bse/ sse
-    real(dp) :: mt,tm,tn,tscls(20),lums(10),GB(10),zpars(20)
-    real(dp) :: mc,rc,menv,renv,k2,mcx,r,lum,epoch
-    
+    ! dummy variables for bse/sse
+    real(dp):: mt, tm, tn, tscls(20), lums(10), GB(10), zpars(20)
+    real(dp):: mc, rc, menv, renv, k2, mcx, r, lum, epoch
+   
     idd = 1
     if(present(id)) idd = id
     t => tarr(idd)
@@ -38,12 +38,12 @@ subroutine evolv_metisse(mass,max_age,ierr,id)
     tm = 0.d0
     tn = 0.d0
     str = int(mass*100)
-    call METISSE_star(t% pars% phase,mass,mt,tm,tn,tscls,lums,GB,zpars,timestep,id)
+    call METISSE_star(t% pars% phase, mass, mt, tm, tn, tscls, lums, GB, zpars, timestep, id)
 
     if (t% complete .eqv..false.) ierr = 1
 
   !write the mass interpolated track if write_eep_file is true
-    if (write_eep_file) call write_eep_track(t,mass)
+    if (write_eep_file) call write_eep_track(t, mass)
 
     lines = 0
     tphys = 0.d0
@@ -57,9 +57,9 @@ subroutine evolv_metisse(mass,max_age,ierr,id)
     if (write_output_to_file) then
         output = .true.
         io = alloc_iounit(ierr)
-        write (output_file,"(a,a,i5.5,a)") trim(METISSE_DIR), "/output/evolve_", str, "M.dat"
-        open (io,FILE=trim(output_file),action="write")
-        write(io,'(9a15,2a10)') "time", "age", "mass","core_mass","He_core" &
+        write (output_file, "(a, a, i5.5, a)") trim(METISSE_DIR), "/output/evolve_", str, "M.dat"
+        open (io, FILE = trim(output_file), action="write")
+        write(io, '(9a15, 2a10)') "time", "age", "mass","core_mass","He_core" &
                     ,"CO_core","log_L","log_Teff","log_radius", "phase","e"
     endif
     
@@ -69,17 +69,17 @@ subroutine evolv_metisse(mass,max_age,ierr,id)
         ! advance the time
         tphys = tphys+timestep
         
-        !evolve the star- calculate stellar parameters at tphys
+        !evolve the star-calculate stellar parameters at tphys
[…]
