+++ src/interp_support.f90
@@ -2,11 +2,11 @@ module interp_support
     use track_support
     implicit none
 
-    integer, parameter :: no_interpolation = 0
-    integer, parameter :: linear = 1
-    integer, parameter :: Steffen1990 = 2
-    !integer, parameter :: extrapolation = 3
-    logical :: debug_mass
+    integer, parameter:: no_interpolation = 0
+    integer, parameter:: linear = 1
+    integer, parameter:: Steffen1990 = 2
+    !integer, parameter:: extrapolation = 3
+    logical:: debug_mass
 
     contains
 
@@ -14,39 +14,39 @@ module interp_support
     subroutine interpolate_mass(t, exclude_core)
         implicit none
 
-        logical, intent(in) :: exclude_core
+        logical, intent(in):: exclude_core
 
-        real(dp) :: mass
-        type(track), pointer :: t
+        real(dp):: mass
+        type(track), pointer:: t
 
-        integer :: iseg, keyword,min_index
-        type(track), pointer :: a(:), s(:)
-        real(dp) :: f(3), dx, x(4), y(4), alfa, beta
-        integer :: i, j, k, mlo, mhi,nt,age_col,start
-        integer, allocatable :: eeps(:), excl_cols(:)
+        integer:: iseg, keyword, min_index
+        type(track), pointer:: a(:), s(:)
+        real(dp):: f(3), dx, x(4), y(4), alfa, beta
+        integer:: i, j, k, mlo, mhi, nt, age_col, start
+        integer, allocatable:: eeps(:), excl_cols(:)
         
         debug_mass = .false.
 !        if(t% is_he_track)debug_mass = .true.
 
-        if (debug_mass) print*, 'in interpolate_mass',t% initial_mass,t% pars% phase
+        if (debug_mass) print*, 'in interpolate_mass',t% initial_mass, t% pars% phase
         mass = t% initial_mass
         
-        if (mass/=mass .or. mass<=0.d0) then
-            write(UNIT=err_unit,fmt=*)"METISSE error: Encountered invalid mass value",mass,t% pars% phase
+        if (mass /= mass .or. mass <= 0.d0) then
+            write(UNIT = err_unit, fmt=*)"METISSE error: Encountered invalid mass value",mass, t% pars% phase
             t% ierr = -1
 !            call stop_code(err_unit)
             return
         endif
         
-        dx=0d0; alfa=0d0; beta=0d0; x=0d0; y=0d0
+        dx = 0d0; alfa = 0d0; beta = 0d0; x = 0d0; y = 0d0
         
         ! this line is to avoid array length problem with multiple calls to fix-track
         if (allocated(t% tr) .and. (.not.exclude_core)) call deallocate_arrays(t)
         if (allocated(t% bounds)) deallocate(t% bounds)
         
         ! takes a set of EEP-tracks and find tracks for interpolation (a)
-        call findtracks_for_interpolation(mass,t% is_he_track,t% bounds,min_index,keyword,iseg)
-        
+        call findtracks_for_interpolation(mass, t% is_he_track, t% bounds, min_index, keyword, iseg)
+
         mlo = 1
         mhi = size(t% bounds)
         
@@ -66,19 +66,19 @@ module interp_support
             eeps = key_eeps
         endif
         
-        k = minloc(a(mlo:mhi)% ntrack,dim=1)
+        k = minloc(a(mlo:mhi)% ntrack, dim = 1)
         t% min_index = min_index
         
-        if (debug_mass) print*,"mass, keyword", mass,keyword,t% is_he_track
+        if (debug_mass) print*,"mass, keyword", mass, keyword, t% is_he_track
         if (debug_mass) print*,"interpolate mass" , a% initial_mass
[…]

[…]
