### 2025-08-13 #7877 @JBorrow
Hi @ddhendriks and @s-goldman, as authors of similar JOSS papers in the past, would you be interested in reviewing this paper and package?
