### 2025-08-14 #7877 @ddhendriks
Hi @JBorrow,

Yes I'd be happy to review METISSE.
