### 2025-08-21 55f7db2
add helper subroutines and use statements for the comsic-metisse interface
