+++ src/METISSE_zcnsts.f90
@@ -1,6 +1,7 @@
 subroutine METISSE_zcnsts(z,zpars,ierr)
     use track_support
     use z_support
+    use c_m_interface
 
     real(dp), intent(in) :: z
     real(dp), intent(out) :: zpars(20)
@@ -114,43 +115,48 @@ subroutine METISSE_zcnsts(z,zpars,ierr)
         ! these file contain information about eep tracks, their metallicity
         ! and the format file
         
-        if (len(trim(METALLICITY_DIR))< 1) then
-            write(*,*) "METISSE error: METALLICITY_DIR/path_to_tracks is an empty string"
-            ierr = 1
-            return
-        else
-            temp_filename = '.Zfilenames_H.txt'
-            call get_metallicity_file_list(METALLICITY_DIR,metallicity_file_list,temp_filename)
-                
-            if (.not. allocated(metallicity_file_list)) then
-                write(*,*) "METISSE error: metallicity file(s) not found in ", trim(METALLICITY_DIR)
-                write(*,*) "check if METALLICITY_DIR/path_to_tracks is correct"
+        select case(front_end)
+        case(COSMIC)
+            ! hold for now since we read eeps directly
+        case default    
+            if (len(trim(METALLICITY_DIR))< 1) then
+                write(*,*) "METISSE error: METALLICITY_DIR/path_to_tracks is an empty string"
                 ierr = 1
                 return
             else
-                if(debug) print*,'metallicity files: ',metallicity_file_list
-                call get_metallicity_list(metallicity_file_list,Z_H)
+                temp_filename = '.Zfilenames_H.txt'
+                call get_metallicity_file_list(METALLICITY_DIR,metallicity_file_list,temp_filename)
+                    
+                if (.not. allocated(metallicity_file_list)) then
+                    write(*,*) "METISSE error: metallicity file(s) not found in ", trim(METALLICITY_DIR)
+                    write(*,*) "check if METALLICITY_DIR/path_to_tracks is correct"
+                    ierr = 1
+                    return
+                else
+                    if(debug) print*,'metallicity files: ',metallicity_file_list
+                    call get_metallicity_list(metallicity_file_list,Z_H)
+                endif
             endif
-        endif
-        
-        if (len(trim(METALLICITY_DIR_HE))< 1) then
-            write(out_unit,*) "Warning: METALLICITY_DIR_HE/path_to_he_tracks is an empty string"
-            write(out_unit,*) "Switching to SSE formulae for helium stars "
-            use_sse_NHe = .true.
-        else
-            temp_filename  = '.Zfilenames_He.txt'
-            call get_metallicity_file_list(METALLICITY_DIR_HE,metallicity_file_list_he,temp_filename)
-
-            if (.not. allocated(metallicity_file_list_he)) then
-                write(*,*) "METISSE error: metallicity file(s) not found in ", trim(METALLICITY_DIR_HE)
-                write(*,*) "check if METALLICITY_DIR_HE/path_to_he_tracks is correct"
-                ierr = 1
-                return
+            
+            if (len(trim(METALLICITY_DIR_HE))< 1) then
+                write(out_unit,*) "Warning: METALLICITY_DIR_HE/path_to_he_tracks is an empty string"
+                write(out_unit,*) "Switching to SSE formulae for helium stars "
+                use_sse_NHe = .true.
             else
-                if(debug) print*,'metallicity files he : ', metallicity_file_list_he
-                call get_metallicity_list(metallicity_file_list_he,Z_He)
+                temp_filename  = '.Zfilenames_He.txt'
+                call get_metallicity_file_list(METALLICITY_DIR_HE,metallicity_file_list_he,temp_filename)
+    
[…]
