+++ src/z_support.f90
@@ -1632,6 +1632,174 @@ module z_support
         zpars(14) = z**0.4d0
 
     end subroutine
+
+        ! --------------------------
+    ! Apply Python-provided values based on track type
+    ! --------------------------
+    subroutine apply_cosmic_format_controls(track_type)
+        use c_m_interface
+        character(len=*), intent(in) :: track_type
+
+        if (track_type == 'H') then
+            read_eep_files = py_read_eep_files_H
+            PreMS_EEP = py_PreMS_EEP_H
+            ZAMS_EEP = py_ZAMS_EEP_H
+            IAMS_EEP = py_IAMS_EEP_H
+            TAMS_EEP = py_TAMS_EEP_H
+            BGB_EEP = py_BGB_EEP_H
+            cHeIgnition_EEP = py_cHeIgnition_EEP_H
+            cHeBurn_EEP = py_cHeBurn_EEP_H
+            TA_cHeB_EEP = py_TA_cHeB_EEP_H
+            TPAGB_EEP = py_TPAGB_EEP_H
+            cCBurn_EEP = py_cCBurn_EEP_H
+            post_AGB_EEP = py_post_AGB_EEP_H
+            Initial_EEP = py_Initial_EEP_H
+            Final_EEP = py_Final_EEP_H
+            fix_track = py_fix_track_H
+            low_mass_final_eep = py_low_mass_final_eep_H
+            high_mass_final_eep = py_high_mass_final_eep_H
+
+            age_colname = py_age_colname_H
+            mass_colname = py_mass_colname_H
+            log_L_colname = py_log_L_colname_H
+            log_T_colname = py_log_T_colname_H
+            log_R_colname = py_log_R_colname_H
+            he_core_mass = py_he_core_mass_H
+            co_core_mass = py_co_core_mass_H
+            he_core_radius = py_he_core_radius_H
+            co_core_radius = py_co_core_radius_H
+            mass_conv_envelope = py_mass_conv_envelope_H
+            radius_conv_envelope = py_radius_conv_envelope_H
+            log_Tc = py_log_Tc_H
+            He4_mass_frac = py_He4_mass_frac_H
+            c12_mass_frac = py_c12_mass_frac_H
+            o16_mass_frac = py_o16_mass_frac_H
+
+        else if (track_type == 'He') then
+            read_eep_files = py_read_eep_files_He
+            BGB_EEP = py_BGB_EEP_He
+            cHeBurn_EEP = py_cHeBurn_EEP_He
+            TA_cHeB_EEP = py_TA_cHeB_EEP_He
+            TPAGB_EEP = py_TPAGB_EEP_He
+            cCBurn_EEP = py_cCBurn_EEP_He
+            post_AGB_EEP = py_post_AGB_EEP_He
+            Initial_EEP = py_Initial_EEP_He
+            Final_EEP = py_Final_EEP_He
+            fix_track = py_fix_track_He
+            low_mass_final_eep = py_low_mass_final_eep_He
+            high_mass_final_eep = py_high_mass_final_eep_He
+
+            age_colname = py_age_colname_He
+            mass_colname = py_mass_colname_He
+            log_L_colname = py_log_L_colname_He
+            log_T_colname = py_log_T_colname_He
+            log_R_colname = py_log_R_colname_He
+            he_core_mass = py_he_core_mass_He
+            co_core_mass = py_co_core_mass_He
+            he_core_radius = py_he_core_radius_He
+            co_core_radius = py_co_core_radius_He
+            mass_conv_envelope = py_mass_conv_envelope_He
+            radius_conv_envelope = py_radius_conv_envelope_He
+            log_Tc = py_log_Tc_He
[…]
