### 2025-08-21 0417950
remove unnecessary allocate and add print debugging
