+++ src/z_support.f90
@@ -31,7 +31,7 @@ module z_support
     real(dp) :: high_mass_limit = 1d1 !Msun
     real(dp) :: he_core_mass_limit = 2.2d0 !Msun
 
-    logical :: debug_z = .false.
+    logical :: debug_z = .true.
 
 
 
@@ -1774,6 +1774,7 @@ module z_support
             xa(i)%neep = neep_arr(i)
             xa(i)%ncol = ncol_arr(i)
             xa(i)%is_he_track = is_he_arr(i)
+            print*, 'initial_mass 1', initial_mass(i), xa(i)% initial_mass
     
             ! Allocate arrays
             allocate(xa(i)%eep(neep_arr(i)))
@@ -1782,23 +1783,43 @@ module z_support
     
             ! Copy data
             xa(i)%eep = eep_data(1:neep_arr(i), i)
+            print*, 'number of tracks', ntrack_arr(i)
             do j = 1, ntrack_arr(i)
-                xa(i)%tr(:, j) = tr_data(1:ncol_arr(i), sum(ntrack_arr(1:i-1)) + j)
+                xa(i)%tr(:, j) = tr_data(1:ncol_arr(i), offset + j)
+                if (xa(i)%is_he_track) print*, 'j: ',j
+                if (xa(i)%is_he_track) print*, xa(i)%tr(:, j)
+                if (xa(i)%is_he_track) print*, '---------------------'
+                if (xa(i)%is_he_track) print*, '---------------------'
             end do
+            
             do j = 1, ncol_arr(i)
                 xa(i)%cols(j)%name = col_names(j, i)
             end do
             !determine column of mass, age etc.
-            if (get_cols) call get_named_columns(xa(i)% cols, xa(i)% ncol,xa(i)% is_he_track)
+            print*, xa(i)% cols
+            call get_named_columns(xa(i)% cols, xa(i)% ncol,xa(i)% is_he_track)
             if (code_error) return
             
             if (xa(i)% is_he_track) then
                 xa(i)% initial_mass = xa(i)% tr(i_mass,ZAMS_HE_EEP)
+                print*, 'initial_mass 2', initial_mass(i), xa(i)% initial_mass
+                print*, 'indices', i_mass, ZAMS_HE_EEP
+                
             else
                 xa(i)% initial_mass = xa(i)% tr(i_mass,ZAMS_EEP)
+                print*, 'initial_mass 2', initial_mass(i), xa(i)% initial_mass
+                print*, 'indices', i_mass, ZAMS_EEP
             endif
+            
             call set_star_type_from_history(xa(i))
+            
+            offset = offset + ntrack_arr(i)
+
         end do  
+
+        
+
+
     end subroutine set_tracks_from_python_inputs
     
     elemental function relative_diff(z1,z2) result(y)
