### 2025-08-22 e749fd2
removed print statements and fixed conflicts
