### 2025-08-22 607c6b2
removed print statements and fixed conflicts
