### 2025-08-22 7c79845
removed print statements and fixed conflicts
