### 2025-08-22 64496d1
add cosmic interface here instead of at the COSMIC repo
