+++ src/c_m_interface.f90
@@ -1,0 +1,331 @@
+module c_m_interface
+    implicit none
+    integer, parameter :: f2py_strlen = 256
+
+    ! -----------------------
+    ! Python track info
+    ! -----------------------
+    ! Arrays
+    integer :: ntracks_h_in, max_ncol_h_in, max_neep_h_in, max_points_h_in
+    logical :: is_he_in
+    character(len=f2py_strlen), allocatable :: filenames_h_in(:)
+    real(8), allocatable :: initial_mass_h_in(:)
+    real(8), allocatable :: initial_Y_h_in(:)
+    real(8), allocatable :: initial_Z_h_in(:)
+    real(8), allocatable :: Fe_div_H_h_in(:)
+    real(8), allocatable :: alpha_div_Fe_h_in(:)
+    real(8), allocatable :: v_div_vcrit_h_in(:)
+    integer, allocatable :: ntrack_arr_h_in(:)
+    integer, allocatable :: neep_arr_h_in(:)
+    integer, allocatable :: ncol_arr_h_in(:)
+    logical, allocatable :: is_he_arr_h_in(:)
+    integer, allocatable :: eep_data_h_in(:,:)       ! (max_neep, ntracks)
+    real(8), allocatable :: tr_data_h_in(:,:)       ! (max_ncol, max_points)
+    character(len=f2py_strlen), allocatable :: col_names_h_in(:,:) ! (max_ncol, ntracks)
+
+    ! Similarly for helium tracks
+    integer :: ntracks_he_in, max_ncol_he_in, max_neep_he_in, max_points_he_in
+    character(len=f2py_strlen), allocatable :: filenames_he_in(:)
+    real(8), allocatable :: initial_mass_he_in(:)
+    real(8), allocatable :: initial_Y_he_in(:)
+    real(8), allocatable :: initial_Z_he_in(:)
+    real(8), allocatable :: Fe_div_H_he_in(:)
+    real(8), allocatable :: alpha_div_Fe_he_in(:)
+    real(8), allocatable :: v_div_vcrit_he_in(:)
+    integer, allocatable :: ntrack_arr_he_in(:)
+    integer, allocatable :: neep_arr_he_in(:)
+    integer, allocatable :: ncol_arr_he_in(:)
+    logical, allocatable :: is_he_arr_he_in(:)
+    integer, allocatable :: eep_data_he_in(:,:)       
+    real(8), allocatable :: tr_data_he_in(:,:)       
+    character(len=f2py_strlen), allocatable :: col_names_he_in(:,:) 
+    ! -----------------------
+    ! File lists
+    ! -----------------------
+    character(len=f2py_strlen), allocatable :: py_track_list(:)
+    character(len=f2py_strlen), allocatable :: py_track_list_he(:)
+    character(len=f2py_strlen), allocatable :: py_metallicity_file_list(:)
+    character(len=f2py_strlen), allocatable :: py_metallicity_file_list_he(:)
+
+    ! -----------------------
+    ! Metallicity values
+    ! -----------------------
+    real(8), allocatable :: py_Z_list(:)
+    real(8), allocatable :: py_Z_list_he(:)
+
+    ! -----------------------
+    ! Hydrogen format controls
+    ! -----------------------
+    logical :: py_read_eep_files_H
+    integer :: py_PreMS_EEP_H, py_ZAMS_EEP_H, py_IAMS_EEP_H, py_TAMS_EEP_H, py_BGB_EEP_H
+    integer :: py_cHeIgnition_EEP_H, py_cHeBurn_EEP_H, py_TA_cHeB_EEP_H, py_TPAGB_EEP_H
[…]
