### 2025-08-22 3d476e8
added a workaround for bh3 tracks
