### 2025-08-22 e9792fe
fixed conflicts for makefile
