### 2025-08-22 b890625
kept changes for amuse-interface
