### 2025-08-23 1525a6a
added c_m_interface to zcnsts in makefile
