### 2025-08-23 611bdb7
Update makefile for c_m_interface
