+++ make/makefile
@@ -16,22 +16,38 @@ FFLAGS2= -ffixed-form -fimplicit-none $(DEBUG)
 # metisse in standalone mode
 METISSE=metisse
 
+SRCE_main = main_metisse.f90
+OBJT_main = $(SRCE_main:.f90=.o)
+
+TEST=test
+
+SRCE_test = test_metisse.f90
+OBJT_test = $(SRCE_test:.f90=.o)
+
 # metisse as a library
 LIBMETISSE=libmetisse
 
-SRCE_main = main_metisse.f90
-
-SRCE_test = test_metisse.f90
-
 SRCE0 = \
-evolv_metisse.f90 assign_commons_main.f90 assign_commons_other.f90 \
-track_support.f90 imf_support.f90 z_support.f90 sse_support.f90 \
-remnant_support.f90 interp_support.f90 \
-METISSE_miscellaneous.f90 METISSE_zcnsts.f90 METISSE_star.f90 \
-METISSE_hrdiag.f90 METISSE_deltat.f90
+		evolv_metisse.f90 \
+		assign_commons_main.f90 \
+		assign_commons_other.f90 \
+		c_m_interface.f90 \
+		track_support.f90 \
+		imf_support.f90 \
+		z_support.f90 \
+		sse_support.f90 \
+		remnant_support.f90 \
+		interp_support.f90 \
+		METISSE_miscellaneous.f90 \
+		METISSE_zcnsts.f90 \
+		METISSE_star.f90 \
+		METISSE_hrdiag.f90 \
+		METISSE_deltat.f90
 
 SRCE00 = \
-assign_remnant.f hrdiag_remnant.f mrenv.f \
+		assign_remnant.f \
+		hrdiag_remnant.f \
+		mrenv.f 
 
 LIBSRCE = \
 		evolv_metisse.f90 \
@@ -65,16 +81,16 @@ LIBSRCE1 = \
 	   evolv_metisse.f90 \
 	   assign_commons_other.f90
 	
-
-OBJT_main = $(SRCE_main:.f90=.o)
-OBJT_test = $(SRCE_test:.f90=.o)
 OBJT0 = $(SRCE0:.f90=.o)
 OBJT00 = $(SRCE00:.f=.o)
 LIBOBJ = $(LIBSRCE:.f90=.o)
 	
 track_support.o: $(METISSE_src_dir)/track_support.f90
+
+c_m_interface.o: $(METISSE_src_dir)/c_m_interface.f90
+
 imf_support.o: $(METISSE_src_dir)/imf_support.f90 track_support.o
-z_support.o: $(METISSE_src_dir)/z_support.f90 track_support.o
+z_support.o: $(METISSE_src_dir)/z_support.f90 track_support.o c_m_interface.o
 interp_support.o: $(METISSE_src_dir)/interp_support.f90 track_support.o
 sse_support.o: $(METISSE_src_dir)/sse_support.f90 track_support.o
 remnant_support.o: $(METISSE_src_dir)/remnant_support.f90 track_support.o sse_support.o
@@ -90,12 +106,11 @@ METISSE_star.o: $(METISSE_src_dir)/METISSE_star.f90 track_support.o interp_suppo
 METISSE_hrdiag.o: $(METISSE_src_dir)/METISSE_hrdiag.f90 track_support.o interp_support.o sse_support.o remnant_support.o
 METISSE_gntage.o: $(METISSE_src_dir)/METISSE_gntage.f90 track_support.o interp_support.o
 
[…]
