### 2025-08-23 ba34716
Update METISSE_zcnsts.f90
