+++ make/makefile
@@ -90,7 +90,7 @@ track_support.o: $(METISSE_src_dir)/track_support.f90
 c_m_interface.o: $(METISSE_src_dir)/c_m_interface.f90
 
 imf_support.o: $(METISSE_src_dir)/imf_support.f90 track_support.o
-z_support.o: $(METISSE_src_dir)/z_support.f90 track_support.o 
+z_support.o: $(METISSE_src_dir)/z_support.f90 track_support.o c_m_interface.o
 interp_support.o: $(METISSE_src_dir)/interp_support.f90 track_support.o
 sse_support.o: $(METISSE_src_dir)/sse_support.f90 track_support.o
 remnant_support.o: $(METISSE_src_dir)/remnant_support.f90 track_support.o sse_support.o
