### 2025-08-23 80e9f82
reverted the CO core check to org
