### 2025-08-23 5879e90
workaround for tracks with decaying co core numerical issue
