### 2025-08-23 e0cc517
remnant mass fix for very low-mass stars with no co core
