### 2025-08-23 46d0c27
remnant mass fix for very low-mass stars with no co core
