### 2025-08-23 c9bd75c
removed the redunant sse remnant conditions check unless for sse_nhe stars
