### 2025-09-02 95500a5
debug and fixes for cosmic python fileio
