+++ src/METISSE_zcnsts.f90
@@ -91,6 +91,10 @@ subroutine METISSE_zcnsts(z, zpars, ierr)
             if (ierr /= 0) call stop_code
         case(COSMIC)
              ! call get_COSMIC_input()
+            if (.not. allocated(filenames_he_in)) then
+                write(out_unit,*)"Switching to SSE formulae for helium stars "
+                use_sse_NHe = .true.
+            endif
         case(AMUSE)
             ! If AMUSE has not set METALLICITY_DIR, it will use the defaults from METISSE
             if (len(trim(amuse_metallicity_dir)) > 0) METALLICITY_DIR = amuse_metallicity_dir
@@ -188,38 +192,34 @@ subroutine METISSE_zcnsts(z, zpars, ierr)
             ! we won't read in any metallicity files
             ! instead we will read in the format files and eeps 
             ! directly with COSMIC and pass them to METISSE
-
             if (i == 2) then
                 ! Naked helium stars
-                if (allocated(py_track_list_he)) then
+                ! if (allocated(py_track_list_he)) then
                     ! First get info on the properties of the track_list
-                    if (allocated(track_list)) deallocate(track_list)
-                    allocate(track_list(size(py_track_list_he)))
-                    track_list = py_track_list_he
-                    USE_DIR = METALLICITY_DIR_HE
-                    num_tracks = size(track_list)
+                    ! if (allocated(track_list)) deallocate(track_list)
+                    ! allocate(track_list(size(py_track_list_he)))
+                    ! track_list = py_track_list_he
+                    ! USE_DIR = METALLICITY_DIR_HE 
                     call apply_cosmic_format_controls('He')
                     call read_key_eeps_he()
                     if (debug) print*, "key eeps for he stars", key_eeps_he
                     call set_tracks_from_python_inputs(.true.)
-                end if
             else
                 ! Hydrogen rich stars
-                if (allocated(py_track_list)) then
+                ! if (allocated(py_track_list)) then
                     ! First get info on the properties of the track_list
-                    if (allocated(track_list)) deallocate(track_list)
-                    allocate(track_list(size(py_track_list)))
-                    track_list = py_track_list
-                    USE_DIR = METALLICITY_DIR
-                    num_tracks = size(track_list)
+                    ! if (allocated(track_list)) deallocate(track_list)
+                    ! allocate(track_list(size(py_track_list)))
+                    ! track_list = py_track_list
+                    ! USE_DIR = METALLICITY_DIR
                     call apply_cosmic_format_controls('H')
                     call read_key_eeps()
                     if (debug) print*, "key eeps", key_eeps   
                     call set_tracks_from_python_inputs(.false.)
-                end if
+                ! end if
             endif
             
-            if (debug) print*, "num_tracks", num_tracks
+            if (debug) print*, "num_tracks", size(xa)
             if (debug) print*, "tracks set by COSMIC"
 
         case default
@@ -295,7 +295,7 @@ subroutine METISSE_zcnsts(z, zpars, ierr)
             if (read_eep_files) then
                 if (debug) print*,"reading eep files"
                 do j=1,num_tracks
-                    call read_eep(xa(j))
+                    call read_MIST_track(xa(j))
                     if (code_error) return
 !                    if(debug) write(*,'(a100,f8.2,99i8)') trim(xa(j)% filename), xa(j)% initial_mass, xa(j)% ncol
                 end do
@@ -319,24 +319,31 @@ subroutine METISSE_zcnsts(z, zpars, ierr)
                 end if
     
                 do j=1,num_tracks
-                    call read_input_file(xa(j))
[…]
