+++ src/z_support.f90
@@ -315,7 +315,7 @@ module z_support
         endif
     end subroutine
     
-    subroutine read_eep(x)
+    subroutine read_MIST_track(x)
         !adapted from iso/iso_eep_support.f90
         type(track), intent(inout) :: x
         integer :: ierr
@@ -386,10 +386,10 @@ module z_support
         close(io)
         call free_iounit(io)
     
-    end subroutine read_eep
+    end subroutine read_MIST_track
     
     !adapted from read_history_file of iso_eep_support
-    subroutine read_input_file(x)
+    subroutine read_other_track(x)
         type(track), intent(inout) :: x
         integer :: ierr
         
@@ -406,6 +406,7 @@ module z_support
         open(unit=io,file=trim(x% filename),status='old',action='read')
         !read lines of header as comments
 
+        ! Handle header
         if (header_location>0)then
             do i = 1,header_location-1
                 read(io,*) !header
@@ -413,10 +414,10 @@ module z_support
             allocate(temp_cols(total_cols))
             !get column names
             read(io,'(a)') line
-            do i =1, total_cols
+            do i = 1, total_cols
                 j = scan(line," ")
                 temp_cols(i)% name = line(1:j)
-                if (trim(temp_cols(i)% name)==extra_char) then
+                if (trim(temp_cols(i)% name)/=extra_char) then
                     line = adjustl(line(j:))
                     j = scan(line," ")
                     temp_cols(i)% name = line(1:j)
@@ -424,7 +425,8 @@ module z_support
                 line = adjustl(line(j:))
             end do
         endif
-        !figure out how many data lines
+
+        ! Count data lines
         j=0
         do while(.true.)
             read(io,*,iostat=ierr)
@@ -442,11 +444,13 @@ module z_support
             enddo
         endif
 
+        ! allocate arrays
         x% ntrack = j
         x% ncol = total_cols
         allocate(x% tr(x% ncol, x% ntrack),x% cols(x% ncol))
         x% cols% name = temp_cols% name
 
+        ! Fill track data
         do j=1, x% ntrack
             read(io,'(a)') line
             call split(line, x% tr(:,j), x% ncol)
@@ -461,6 +465,7 @@ module z_support
         
         if(header_location >0) deallocate(temp_cols)
 
+        ! Setup eep arrays
         if (x% is_he_track) then
             x% neep = count(key_eeps_he .le. x% ntrack,1)
             allocate(x% eep(x% neep))
@@ -471,18 +476,20 @@ module z_support
             x% eep = pack(key_eeps,mask = key_eeps .le. x% ntrack)
         endif
 
+        ! Initial mass
         if (x% is_he_track) then
             x% initial_mass = x% tr(i_mass,ZAMS_HE_EEP)
         else
             x% initial_mass = x% tr(i_mass,ZAMS_EEP)
         endif
         
+        ! Other initial params
         call set_star_type_from_history(x)
         x% initial_Z = initial_Z
         x% initial_Y = Y_files
 
         if (debug) print*,x% initial_mass, x% initial_Z, x% ncol
-    end subroutine read_input_file
+    end subroutine read_other_track
 
     !from C.Flynn's driver routine
 
@@ -544,10 +551,7 @@ module z_support
         i_co_core = locate_column(cols, co_core_mass, essential)
         
[…]
