### 2025-09-02 02a47c1
debug and fixes for cosmic python fileio
