### 2025-09-02 c8d44d2
removed extra prints
