### 2025-09-16 ed83454
fixed core mass error for very low mass stars
