+++ src/remnant_support.f90
@@ -24,89 +24,70 @@
     integer :: wd_flag = 0
     integer :: ec_flag = 0
     integer :: if_flag = 0
-
-    logical :: end_of_file
     logical :: debug_rem = .false.
 
     contains
-    
-    logical function check_remnant_phase(pars)
-        type(star_parameters) :: pars
 
-        check_remnant_phase = .true.
+    subroutine check_early_end(t,dt_hold,id)
+        real(dp) :: dt_hold
+        type(track), pointer :: t
+        integer :: id
 
-        pars% core_mass = pars% McCO
-        if ((pars% phase >0) .and. (pars% core_mass<tiny)) then
-            write(UNIT=err_unit,fmt=*)"METISSE error: non-positive core mass",pars% core_mass
-            if (front_end/=COSMIC) code_error = .true.
-            !assigning an ad-hoc non-zero core mass so the code doesn't break
-            pars% core_mass = 0.75*pars% McHe
-        endif
-        pars% age_old = pars% age
-        if (debug_rem) print*, 'Remnant: mass, core_mass ', pars% mass, pars% McHe, pars% McCO
-        
-        end function check_remnant_phase
-        
-        subroutine check_early_end(t,dt_hold,id)
-            real(dp) :: dt_hold
-            type(track), pointer :: t
-            integer :: id
+        if (t% ierr/=0) return
 
-            if (t% ierr/=0) return
-
-            if ((t% initial_mass.gt.very_low_mass_limit).and. (dt_hold.le.t% pars% dt).and.(t% reju.eqv. .false.)) then
-                write(UNIT=err_unit,fmt=*) 'WARNING: Early end of file at phase, mass and id',&
-                t% pars% phase,t% initial_mass,id,t% reju
-                t% ierr = -1
+        if ((t% initial_mass.gt.very_low_mass_limit).and. (dt_hold.le.t% pars% dt).and.(t% reju.eqv. .false.)) then
+            write(UNIT=err_unit,fmt=*) 'WARNING: Early end of file at phase, mass and id',&
+            t% pars% phase,t% initial_mass,id,t% reju
+            t% ierr = -1
 !                    call stop_code
-            endif
-                
-        end subroutine check_early_end
-                
-        
-        subroutine assign_remnant_METISSE(pars, mcbagb)
-            implicit none
-            type(star_parameters) :: pars
-            real(dp) :: Mcbagb
+        endif
+            
+    end subroutine check_early_end
+            
+    
+    subroutine assign_remnant_METISSE(pars, mcbagb)
+        implicit none
+        type(star_parameters) :: pars
+        real(dp) :: Mcbagb
 
-            !Mup_core, Mec_core are calculated in set_zparmeters routine of zfuncs
-            if(pars% core_mass < M_ch)then
-                if(mcbagb< Mup_core)then
-                    pars% phase = CO_WD        !Zero-age Carbon/Oxygen White Dwarf
+        !Mup_core, Mec_core are calculated in set_zparmeters routine of zfuncs
+        if(pars% core_mass < M_ch)then
+            if(mcbagb< Mup_core)then
+                pars% phase = CO_WD        !Zero-age Carbon/Oxygen White Dwarf
+            else
+                if (ec_flag>0 .and. pars% McCO >= 1.372)then
+                    !electron-capture collapse of an ONe core
+                    !1.372<= mc< 1.44
+                    pars% phase = NS
+                    call initialize_ECSNe(pars)
+                    if (debug_rem) print*,"ECSNe I: Mc< Mch, Mcbagb>Mup"
                 else
-                    if (ec_flag>0 .and. pars% McCO >= 1.372)then
-                     !electron-capture collapse of an ONe core
-                     !1.372<= mc< 1.44
-                        pars% phase = NS
-                        call initialize_ECSNe(pars)
[…]
