### 2025-09-28 #8817 @ddhendriks
Hi @poojanagrawal and others,

Just an update from my side: September was a bit more busy than I expected so It took me a while to get around to this. I have gone through most of the code/docs/paper now but I will have to let is simmer overnight and will complete it early next week.

Kind regards,

David

I
