### 2025-10-02 TeamMETISSE/METISSE#76 @ddhendriks
JOSS REVIEW Documentation section: A statement of need
Revise the “statement of need”:
- [x] Clarify the target audience.
- [x] Describe METISSE’s role as an interpolation engine rather than a full population synthesis tool.
- [x] Make its integration with binary evolution codes a central message in both the documentation and paper.
