### 2025-10-02 TeamMETISSE/METISSE#77 @ddhendriks
JOSS REVIEW Documentation section: Example usage
Extend the “how to use” section:
- [x] Add a practical guide to integrating METISSE into an existing population code, showing the key functions and workflow, and explaining each part.
