### 2025-10-02 TeamMETISSE/METISSE#78 @ddhendriks
JOSS REVIEW Documentation section: Functionality documentation
Expand the functionality docs to:
- [x] Clarify which layers of the code (e.g. interpolation vs. winds/remnants) can be used independently.
- [x] Add a worked “integration” example (pseudocode or minimal Fortran) showing required API calls.
- [x] Supplement the flow diagram with an integration-focused version.
- [x] Document available interpolation methods and their tradeoffs.
