### 2025-10-02 TeamMETISSE/METISSE#79 @ddhendriks
JOSS REVIEW Paper section: Summary
Revise the summary to
  - [x] Be more clear about the integration efforts
  - [x] Expand slightly to highlight the package’s purpose and strengths,
  - [x] Include a direct link to the documentation.
