### 2025-10-02 TeamMETISSE/METISSE#80 @ddhendriks
JOSS REVIEW Paper section: A statement of need
Clarify in the Statement of Need:
  - [x] Whether METISSE is primarily an interpolation engine or a intended as a full stellar evolution framework.
  - [x] Explicitly state whether it includes its own (example) stellar tracks
  - [x] Refine the wording about “enabling systematic studies” to better reflect the role METISSE plays.
