### 2025-10-02 TeamMETISSE/METISSE#81 @ddhendriks
JOSS REVIEW Paper section: State of the field
# action
  Expand the “state of the field” section to
  - [x] Clarify the main distinctions between METISSE and SEVN/ComBinE
  - [x] Briefly situate METISSE relative to other relevant projects (e.g. POSYDON, MINT) to give readers a fuller picture of where it fits.
