### 2025-10-02 TeamMETISSE/METISSE#82 @ddhendriks
JOSS REVIEW Paper section: references
# action
  Expand references to:
  - [x] Include reference/link to the published Zenodo example tracks
