### 2025-10-02 #8817 @ddhendriks
# Review report METISSE
Dear @poojanagrawal, others

Thank you for your patience and for having developed this code. I think it could be a useful tool for the field. After having gone through the code, the documentation, and the paper, I am a bit confused about the core purpose of this code, however.

Seeing that @TomWagg  focussed on the software side (installation, usage), and has already provided his round of comments with a bit of a focus on the functionality, I will not focus on the code actually running. Once the fixes by @TomWagg are processed, I can have a look at that part, if so desired.

My focus has been a bit more on the documentation and the paper, and I do have some comments here. In the section below I cover these comments as replies on the review criteria, and I omit any review criterium that is covered to my satisfaction.

## documentation comments
### *A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?*

The paper describes what the code does and how it can be used, but it’s not yet clear who the intended audience is or how it fits into the larger context.

Currently, the package is presented as handling stellar populations, but support for that seems minimal (single stars, limited sampling distributions, single-core use). This makes it more accurate to describe METISSE as a tool for evolving stellar systems via track interpolation, which can then be integrated into population synthesis codes.

In fact, the documentation already states that it is “specifically designed to be integrated with binary evolution codes”, I think this should be emphasized much more prominently, both in the JOSS paper and on the landing page. Highlighting this as the main purpose would set clearer expectations and showcase the real strength of the package.

- [ ] [Suggested actions for Documentation: A statement of need](https://github.com/TeamMETISSE/METISSE/issues/76)

###  *Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).*

  The examples for single stars and binary stars are helpful. For the binary case, however, what’s missing is guidance on how to perform the interfacing of METISSE with population synthesis codes that currently rely on SSE fitting formulae.

  While users could dig through the linked GitHub code, it would be much more accessible to have a step-by-step explanation in the documentation: which functions need to be called, how to set up the interface, and what the workflow looks like. Since METISSE is specifically designed for integration with binary evolution codes, a dedicated user guide for this integration feels essential.

- [ ] [Suggested actions for Documentation: Example usage](https://github.com/TeamMETISSE/METISSE/issues/77)

### *Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?*

  The documentation is helpful (e.g. the code structure section and flow diagram), but it doesn’t fully support METISSE’s main purpose: integration with contemporary binary evolution codes.

  Right now, it’s unclear how flexible the API is. For example:
  - Can one use only the interpolation layer without the built-in wind or remnant-mass schemes?
  - Is the entire evolve subroutine interchangeable, or do users need to use the full stack?

  While these answers can be found by reading the Fortran code, it would be much clearer (and more approachable) to show this explicitly in the docs. A minimal example Fortran adapter, or pseudocode illustrating which API calls are needed for integration, would go a long way. This comes back to my comment above.

  In addition, the flow diagram is useful but seems tailored to the standalone version. A complementary diagram for the “integration” case would be very valuable, or at least a clearer indication of what part is 'front-end dependent' and can actually be replaced.

  Finally, the interpolation methods themselves deserve more explanation: what approaches are implemented, how they differ, and why a user might choose one over another. This context could help users decide whether METISSE suits their needs.

- [ ] [Suggested actions for the Documentation: Functionality documentation](https://github.com/TeamMETISSE/METISSE/issues/78)

## Software paper comments

First of all, I'd like to suggest to add something like the following strong and clear sentence (taken from the documentation) to the summary of the paper.

``` ...[METISSE] has been specifically designed to be integrated with binary evolution codes that can [evolve stars in binary systems], enabling them to make use of modern stellar data for binary evolution computations.  ```

as I think in eventhough the summary mentions use with other codes, saying that it is specifically designed to do so would put a bit more confidence in that.

### *Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?*

  The high-level purpose of METISSE is described, but the current summary could be clearer and more balanced. For example, the claim that it can be “seamlessly integrated with any population synthesis code using SSE fitting formulae” feels too strong. It is probably easier for Fortran-based codes, but there are other cods that use the SSE fitting formulae without using FORTRAN (e.g. COMPAS, binary_c). I wonder how how easy it then is to apply your framework.

  The summary also feels a bit short. Since this is often the first thing readers see, it would help to expand it slightly, emphasizing the main strengths and intended use cases. Finally, a forward reference to the documentation site would help as well.

- [ ] [Suggested actions for Paper: Summary](https://github.com/TeamMETISSE/METISSE/issues/79)

### *A statement of need: Does the paper have a section titled ‘Statement of need’ that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?*

  The section is present, but it’s not fully clear whether METISSE should be thought of as (a) an interpolation library, or (b) a standalone stellar evolution code. The text suggests both stand-alone use and use as an interface layer, but the distinction could be clearer, and preferably with the focus on the interfacing.

  As the code can be used as a stand-alone version that evolves stars based on pre-calculated tracks, then it would be helpful to point directly to those tracks. If not, it would be useful to make explicit that users are expected to provide their own.

  Finally, the statement that METISSE “enables systematic studies of stellar parameters” could be refined — METISSE makes such studies more accurate and flexible by allowing variation of the global physical assumptions in the stellar tracks, but it does not itself “enables systematic studies of stellar parameters”, as other codes can already vary stellar parameters..

- [ ] [Suggested actions for Paper: A statement of need](https://github.com/TeamMETISSE/METISSE/issues/80)

### *State of the field: Do the authors describe how this software compares to other commonly-used packages?*

  The comparison to other packages is only briefly touched on, and could be strengthened. The summary mentions SEVN and ComBinE, but the actual distinctions are not clear. Both of those also interpolate on detailed stellar tracks, so it would help to spell out more explicitly what sets METISSE apart beyond “seamless integration.” Is it flexibility, modularity, ease of use, or something else?

  It would also be valuable to at least acknowledge other ongoing efforts, such as [POSYDON](https://ui.adsabs.harvard.edu/abs/2023ApJS..264...43F/abstract)) and [MINT](https://ui.adsabs.harvard.edu/abs/2025arXiv250317772R/abstract) (disclaimer: I am a MINT developer), which take somewhat different approaches.

  Moreover if i am not mistaken TRILEGAL (https://www.aanda.org/articles/aa/pdf/2005/24/aa2352-04.pdf), as well as modules BINPSE/BINaPSE https://academic.oup.com/mnras/article/506/4/5681/6321175) also performs some form of interpolation on stellar tracks, although I am not a 100% sure. Might be worth the mention but I understand this is not a review paper.

- [ ] [Suggested actions for Paper: state of the field](https://github.com/TeamMETISSE/METISSE/issues/81)

### *References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?*

  References are fine. Except for my point of explicitly pointing to the example
  detailed stellar tracks as a data source.

- [ ] [Suggested actions for Paper: references](https://github.com/TeamMETISSE/METISSE/issues/82)
