### 2025-10-21 1cc5cce
made number_of_tracks real to allow for exp like 1d+4
