### 2025-10-21 2022488
fixed link to cosmic in documentation
