### 2025-10-21 bda94dc
added copy button to codes
