+++ docs/build/html/_sources/usage_other.md.txt
@@ -84,10 +84,11 @@ The paths provided above are just examples. Users should provide `path_to_tracks
 4. Pass `SSEDict` along with other arguments to the `evolve` function of COSMIC and you are all set. For instance, if you are calculating the evolution of a single binary, the evolve function should look like this: 
 
 ```ipython 
-bpp, bcm, initC, kick_info = Evolve.evolve(initialbinarytable=single_binary,
-                                          BSEDict=BSEDict, 
-                                          SSEDict=SSEDict
-                                          ) 
+bpp, bcm, initC, kick_info = Evolve.evolve(
+                              initialbinarytable=single_binary,
+                              BSEDict=BSEDict, 
+                              SSEDict=SSEDict
+                              ) 
 ``` 
 
 COSMIC will then evolve the system using the provided stellar tracks and store the output in its usual `bpp` and `bcm` arrays.
@@ -120,17 +121,15 @@ $ ./mk
 
 4. When using BSE, METISSE-specific inputs are read through the `METISSE_input_controls` inlist in the *evolve_metisse.in* file. We supply the location of the folders containing [metallicity files](glossary.md#metallicity-file) for hydrogen and helium stars through `METALLICITY_DIR` and `METALLICITY_DIR_HE`. (Make sure to provide relevant paths based on the location of these folders **on your machine** ). 
 
-```
+``` fortran
 &METISSE_input_controls
-
 
 METALLICITY_DIR = '/Users/poojan/Downloads/sample_tracks_solarZ/Hydrogen/'
             
 METALLICITY_DIR_HE = '/Users/poojan/Downloads/sample_tracks_solarZ/Helium/'
 
 /
-
-```
+ ```
 
 All other inputs including the binary parameters as well as the values of mass and metallicity are read through the BSE input file `binary.in`. 
 
@@ -144,23 +143,24 @@ $ ./bse_metisse
 
 6. BSE will produce the following terminal output.
 
-``` console
-      TIME      M1       M2   K1 K2        SEP    ECC  R1/ROL1 R2/ROL2  TYPE
-     0.0000    2.500    1.000  1  1      137.607  0.00   0.028   0.022  INITIAL 
-   567.8636    2.499    1.000  2  1      137.649  0.00   0.063   0.022  KW CHNGE
-   574.5686    2.499    1.000  3  1      137.652  0.00   0.120   0.022  KW CHNGE
-   581.1684    2.499    1.000  4  1      136.123  0.00   0.525   0.022  KW CHNGE
-   768.8467    2.496    1.000  5  1      136.210  0.00   0.293   0.022  KW CHNGE
-   780.5401    2.495    1.000  5  1      125.450  0.00   1.001   0.024  BEG RCHE
-   780.5401    0.523    1.000  8  1       11.851  0.00   1.001   0.024  COMENV  
-   780.5401    0.523    1.000  8  1       11.851  0.00   0.018   0.177  END RCHE
-   781.1571    0.468    1.000 11  1       11.883  0.00   0.004   0.172  KW CHNGE
-  9470.1367    0.468    1.000 11  2       14.083  0.00   0.003   0.208  KW CHNGE
- 11192.3223    0.468    0.999 11  3       14.024  0.00   0.003   0.279  KW CHNGE
- 11950.0928    0.468    0.999 11  3       11.966  0.00   0.004   1.000  BEG RCHE
- 11950.0928    0.468    0.200 11 10        1.199  0.00   0.004   1.000  COMENV  
[…]
