### 2025-10-21 b4f2b82
changed guidelines to FAQ and adpted accordingly
