### 2025-10-21 fe64840
fixed bse_bits_in doc
