+++ docs/build/html/_sources/usage_other.md.txt
@@ -107,19 +107,25 @@ To run METISSE with BSE and compute the evolution of a single binary, say a 2.5 
 $ git clone https://github.com/poojanagrawal/BSE-METISSE
 
 ```
+Get the METISSE submodule.
 
+``` console
 
-2. Make sure use_SSE is false in `bse.f` in the `src` folder.
+$ git submodule update --init --recursive
 
-3. Next, compile the code:
+```
+
+2. Compile the code using:
 
 ``` console
 
 $ ./mk
 
 ```
+<!-- Make sure use_SSE is false in `bse.f` in the `src` folder before compiling. -->
 
-4. When using BSE, METISSE-specific inputs are read through the `METISSE_input_controls` inlist in the *evolve_metisse.in* file. We supply the location of the folders containing [metallicity files](glossary.md#metallicity-file) for hydrogen and helium stars through `METALLICITY_DIR` and `METALLICITY_DIR_HE`. (Make sure to provide relevant paths based on the location of these folders **on your machine** ). 
+
+3. When using BSE, METISSE-specific inputs are read through the `METISSE_input_controls` inlist in the *evolve_metisse.in* file. We supply the location of the folders containing [metallicity files](glossary.md#metallicity-file) for hydrogen and helium stars through `METALLICITY_DIR` and `METALLICITY_DIR_HE`. (Make sure to provide relevant paths based on the location of these folders **on your machine** ). 
 
 ``` fortran
 &METISSE_input_controls
@@ -129,11 +135,52 @@ METALLICITY_DIR = '/Users/poojan/Downloads/sample_tracks_solarZ/Hydrogen/'
 METALLICITY_DIR_HE = '/Users/poojan/Downloads/sample_tracks_solarZ/Helium/'
 
 /
- ```
+ 
+```
 
 All other inputs including the binary parameters as well as the values of mass and metallicity are read through the BSE input file `binary.in`. 
 
-5.  Finally, run BSE using:
+A typical `binary.in` looks like:
+
+
+```{code-block} fortran
+2.5  1.0 13700 100  1 1 0.02 0.0     ! mass1 mass2 maximum_evolution_time orbital_period_in_days initial_type1 initial_type2 metallicity eccentricity
+0.5 0.0 1.0 3.0 0.5                   ! reimers_eta binary_wind_enhancement helium_star_mass_loss_factor alpha_common_envelope lambda_common_envelope
+0 1 0 1 0 2 3.0 29769                 ! common_envelope_flag tides_on_off white_dwarf_initial_final_mass_relation white_dwarf_cooling black_hole_kicks neutron_star_remnant_mass max_neutron_star_mass kick_random_seed
+0.001 0.01 0.02                        ! timestep_pts1 timesteps_pts2 timestep_pts3
+0.0 0.125 1.0 1.5 0.001 10.0 -1.0     ! kick_velocity_dispersion beta_wind_velocity xi_wind_accretion bondi_hoyle_wind_accretion_factor fraction_of_accreted_matter_retained_in_nova_eruption eddington_limit_factor gamma_angular_momentum_loss
+1                                      ! sse_flag
+
+```
+
+<!-- ```{code-block}
+2.5  1.0 13700 100  1 1 0.02 0.0
+0.5 0.0 1.0 3.0 0.5
+0 1 0 1 0 2 3.0 29769
+0.001 0.01 0.02
+0.0 0.125 1.0 1.5 0.001 10.0 -1.0
+1
+```
[…]
