+++ docs/build/html/_sources/usage_standalone.md.txt
@@ -110,7 +110,7 @@ Since we had set, `write_output_to_file = .true.` in `SSE_input_controls`, a [SS
 To compute the evolution of a single stellar population containing, for example, 10,000 stars with initial masses uniformly distributed between 1 and 100 M<sub>$_\odot$</sub>, we modify the `SSE_input_controls` from the single star case as shown below. 
 
 
-```
+``` fortran
 &SSE_input_controls
 
 initial_Z = 0.02
@@ -124,7 +124,7 @@ max_mass = 100.d0
 
 write_track_to_file = .true.   
 /
-
+ 
 ```
 
 We also set `verbose = .false.` in metisse.input to avoid flooding our screen with evolution details of 10,000 stars. 
@@ -144,7 +144,7 @@ To compute a population of single stars with any other distribution of initial m
 If the masses are listed in a file called `my_custom_distribution.txt`, then `SSE_input_controls` will look like this:
 
 
-```
+``` fortran
 &SSE_input_controls
 
 initial_Z = 0.02
@@ -160,7 +160,7 @@ input_mass_file = 'my_custom_distribution.txt'
 write_track_to_file = .true.  
 
 /
-
+ 
 ```
 
 Note that we still need to define `number_of_tracks` as METISSE will read `input_mass_file` for that many stars and will raise errors if the file contains fewer stars.  
@@ -191,7 +191,7 @@ Another option is that we can ask METISSE to use nearby metallicity values.
 This is achieved by increasing `Z_accuracy_limit` in `METISSE_input_controls`.
 METISSE lists available metallicity values in the grid, whenever it cannot find the input value. 
 
-In the sample set, we have tracks for only one metallicity. 
+In the example set, we have tracks for only one metallicity. 
 However, if we were working with a larger grid of stellar tracks, with metallicity values normally distributed between 10<sup>-5</sup> to 10<sup>-1</sup>, we could increase `Z_accuracy_limit` to `8d-2`. This would allow METISSE to run using the closest metallicity in the list, which is `0.0148`. 
 
 ``` console
