+++ docs/source/usage_other.md
@@ -68,10 +68,11 @@ $ pip install .
 The usual input namelist `METISSE_input_controls` is not read when using COSMIC. Instead a Python dictionary, `SSEDict` is used to set the `stellar_engine` and provide the location of the folders containing [metallicity files](glossary.md#metallicity-file) for hydrogen and helium stars through variables `path_to_tracks` and `path_to_he_tracks`.
 
 ``` ipython
-SSEDict = {'stellar_engine': 'metisse', 
-               'path_to_tracks': '/Users/poojan/Downloads/sample_tracks_solarZ/Hydrogen/', 
-               'path_to_he_tracks': '/Users/poojan/Downloads/sample_tracks_solarZ/Helium/' 
-               }
+SSEDict = {
+          'stellar_engine': 'metisse', 
+          'path_to_tracks': '/Users/poojan/Downloads/sample_tracks_solarZ/Hydrogen/', 
+          'path_to_he_tracks': '/Users/poojan/Downloads/sample_tracks_solarZ/Helium/' 
+          }
 ```
 
 :::{Important}
@@ -107,19 +108,25 @@ To run METISSE with BSE and compute the evolution of a single binary, say a 2.5 
 $ git clone https://github.com/poojanagrawal/BSE-METISSE
 
 ```
+Get the METISSE submodule.
 
+``` console
 
-2. Make sure use_SSE is false in `bse.f` in the `src` folder.
+$ git submodule update --init --recursive
 
-3. Next, compile the code:
+```
+
+2. Compile the code using:
 
 ``` console
 
 $ ./mk
 
 ```
+<!-- Make sure use_SSE is false in `bse.f` in the `src` folder before compiling. -->
 
-4. When using BSE, METISSE-specific inputs are read through the `METISSE_input_controls` inlist in the *evolve_metisse.in* file. We supply the location of the folders containing [metallicity files](glossary.md#metallicity-file) for hydrogen and helium stars through `METALLICITY_DIR` and `METALLICITY_DIR_HE`. (Make sure to provide relevant paths based on the location of these folders **on your machine** ). 
+
+3. When using BSE, METISSE-specific inputs are read through the `METISSE_input_controls` inlist in the *evolve_metisse.in* file. We supply the location of the folders containing [metallicity files](glossary.md#metallicity-file) for hydrogen and helium stars through `METALLICITY_DIR` and `METALLICITY_DIR_HE`. (Make sure to provide relevant paths based on the location of these folders **on your machine** ). 
 
 ``` fortran
 &METISSE_input_controls
@@ -134,7 +141,47 @@ METALLICITY_DIR_HE = '/Users/poojan/Downloads/sample_tracks_solarZ/Helium/'
 
 All other inputs including the binary parameters as well as the values of mass and metallicity are read through the BSE input file `binary.in`. 
 
-5.  Finally, run BSE using:
+A typical `binary.in` looks like:
+
+
+```{code-block} fortran
+2.5  1.0 13700 100  1 1 0.02 0.0     ! mass1 mass2 maximum_evolution_time orbital_period_in_days initial_type1 initial_type2 metallicity eccentricity
+0.5 0.0 1.0 3.0 0.5                   ! reimers_eta binary_wind_enhancement helium_star_mass_loss_factor alpha_common_envelope lambda_common_envelope
[…]
