### 2025-10-21 9b85b20
added the forgotten verbose
