### 2025-10-22 f9b8f2d
detailed installation instructions
