### 2025-10-29 f194689
modified doc and paper to higlight usage and integration
