+++ paper.bib
@@ -187,3 +187,39 @@ archivePrefix = {arXiv},
 }    
 
 }
+
+@ARTICLE{Girardi:2005,
+       author = {{Girardi}, L. and {Groenewegen}, M.~A.~T. and {Hatziminaoglou}, E. and {da Costa}, L.},
+        title = "{Star counts in the Galaxy. Simulating from very deep to very shallow photometric surveys with the TRILEGAL code}",
+      journal = {\aap},
+     keywords = {stars: luminosity function, mass function, stars: Hertzsprung-Russell (HR) and C-M diagrams, surveys, Galaxy: stellar content, Galaxy: structure, Astrophysics},
+         year = 2005,
+        month = jun,
+       volume = {436},
+       number = {3},
+        pages = {895-915},
+          doi = {10.1051/0004-6361:20042352},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0504047},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2005A&A...436..895G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Rees:2025,
+       author = {{Rees}, Natalie R. and {Izzard}, Robert G. and {Hendriks}, David D.},
+        title = "{A stellar evolutionary grid for binary population synthesis: from the main sequence to helium ignition}",
+      journal = {arXiv e-prints},
+     keywords = {Solar and Stellar Astrophysics},
+         year = 2025,
+        month = mar,
+          eid = {arXiv:2503.17772},
+        pages = {arXiv:2503.17772},
+          doi = {10.48550/arXiv.2503.17772},
+archivePrefix = {arXiv},
+       eprint = {2503.17772},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025arXiv250317772R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
