+++ paper.md
@@ -10,7 +10,7 @@ authors:
   - name: Poojan Agrawal
     orcid: 0000-0002-1135-984X
     affiliation: 1 
-  - name: Katie Brievik
+  - name: Katelyn Breivik
     orcid: 0000-0001-5228-6598
     affiliation: 3 
   - name: Jarrod Hurley
@@ -53,7 +53,7 @@ bibliography: paper.bib
 
 # Summary
 
-METISSE is an open-source code that interpolates between pre-computed stellar models to quickly derive stellar parameters for population synthesis codes. Written in Modern Fortran, METISSE is both fast and robust. It is comparable to existing rapid stellar evolution codes, such as SEVN [@Iorio:2022] and Combine [@Kruckow:2018] but with the added advantage that it can be seamlessly integrated with any population synthesis code that currently uses popular SSE fitting formulae [@Hurley:2000] to calculate stellar parameters. METISSE can function as a standalone code for simulating single stellar populations as well as in conjunction with population synthesis codes for modelling binary stars and star clusters. 
+METISSE is an open-source stellar evolution tool specifically designed to be integrated with binary evolution and population synthesis codes. Unlike traditional rapid stellar evolution schemes based on static fitting formulae (e.g. SSE; [@Hurley:2000]), METISSE interpolates between pre-computed stellar models to quickly derive stellar parameters for population synthesis codes. This approach makes it straightforward to incorporate the latest stellar physics and explore the effects of different input models. METISSE can be seamlessly integrated with any population synthesis code that currently uses popular Fortran code SSE [@Hurley:2000] to calculate stellar parameters. While METISSE can also be used in standalone mode to evolve single stellar populations, its primary purpose is to act as an interpolation engine within existing population synthesis codes for modelling binary stars and star clusters. Written in Modern Fortran, METISSE is both fast and robust, making it well-suited for large-scale population studies that require realistic stellar evolution input. The full documentation for METISSE is available at []https://metisse.readthedocs.io.
 
 # Statement of need
 
@@ -61,14 +61,17 @@ Stars, especially those with masses greater than eight solar masses (massive sta
 
 As a result, rapid stellar evolution codes that rely on fitting formulas, manually calibrated to specific sets of stellar models to determine the evolution of individual star systems, have been a popular alternative in the past. These methods provide a computationally efficient, fast, and reliable way to calculate stellar population properties. Unfortunately, fitting formulas must be recalculated manually each time for different stellar model sets, limiting our ability to conduct systematic studies of stellar parameters. The use of interpolation in METISSE allows input stellar models to be easily swapped and enables systematic studies of stellar parameters on the stellar populations. 
 
+In recent years, interpolation-based rapid stellar-evolution codes such as TRILEGAL [@Girardi:2005], ComBinE [@Kruckow:2018], SEVN [@Iorio:2022] and MINT [@Rees:2025], as well as machine-learning–based frameworks, have become increasingly common.
+METISSE offers similar speed and purpose but is written in Modern Fortran with full Fortran 77 compatibility and provides analogous SSE subroutines for easy integration into existing population synthesis codes.
+It thus bridges the gap between fast analytic prescriptions and computationally intensive approaches such as self-consistent stellar-structure or machine-learning–based evolution models (e.g., POSYDON), offering a balance of speed, accuracy, and interoperability.
+<!-- These features make METISSE a flexible and efficient option for population synthesis, offering higher physical accuracy than analytic prescriptions without the heavy computational cost of full stellar-structure calculations. -->
 
[…]
