+++ docs/source/structure.md
@@ -1,14 +1,8 @@
 # Code structure
 
+METISSE is designed as an modern alternative to Fortran 77 based SSE [(Hurley et al. 2000)](https://ui.adsabs.harvard.edu/abs/2000MNRAS.315..543H/abstract) for use in population synthesis codes. Its source code, located in the *src* directory, can be broadly divided into two categories: **subroutines analogous to SSE** and **Fortran modules** that provide supporting functions and utilities for METISSE’s data structures and operations.
 
-METISSE has been written completely using Modern Fortran. 
-The source code contains two types of files located in the *src* directory: 
-
-
-## SSE specific subroutines 
-
-
-METISSE has been developed as an alternative to SSE [(Hurley et al. 2000)](https://ui.adsabs.harvard.edu/abs/2000MNRAS.315..543H/abstract) code and, therefore, contains similar subroutines as SSE. The following subroutines replicate the behaviour of the SSE subroutines externally, with similar names and input/output variables.
+## SSE-like subroutines 
 
 | File Name           | Description                                                                                             |
 |---------------------|---------------------------------------------------------------------------------------------------|
@@ -17,13 +11,10 @@ METISSE has been developed as an alternative to SSE [(Hurley et al. 2000)](https
 | METISSE_hrdiag.f90    | Interpolate within the new track to determine stellar parameters at a given age. <br>Also, compute the evolutionary phase of the star including the remnant phases and their properties. |
 | METISSE_deltat.f90    | Calculate the time steps depending on the stage of evolution.                                   |
 | METISSE_mlwind.f90    | Derive the mass loss through stellar winds.                                                    |
-| METISSE_gntage.f90    | Determine the age of a giant star after merger or rejuvenation.                                 |
+| METISSE_gntage.f90    | Determine the age of a giant star after merger  <br> or rejuvenation in a binary interaction.                                 |
 
 
-
-## Modules and other files
-
-The following Fortran modules contain supporting data structures and utilities. 
+## Fortran Modules
 
 | File Name           | Description                                                                                             |
 |---------------------|---------------------------------------------------------------------------------------------------|
@@ -34,38 +25,174 @@ The following Fortran modules contain supporting data structures and utilities.
 | sse_support.f90     | Contains subroutines to calculate SSE-specific quantities.                                         |
 
 
-
-In addition to the above modules, there is an additional file called *METISSE_miscellaneous.f90*. This file contains miscellaneous subroutines that can be called by the binary codes. Ideally, these subroutines should also be organized into a module. However, this is not feasible because the overlying binary evolution codes often have segments written in Fortran 77, and Fortran 77 does not support the use of modules.
-
+In addition to the above modules, there is an additional file called *METISSE_miscellaneous.f90* which contains miscellaneous subroutines that can be called by the binary codes. 
+Although these routines could be organized within a module, they are kept separate to ensure compatibility with legacy Fortran 77 codes, which does not support modules.
 It contains:  
+
 | File Name           | Description                                                                                             |
 |---------------------|---------------------------------------------------------------------------------------------------------|
-| alloc_track.f90     | Allocate the track object.                                                                              |
-| dealloc_track.f90   | Deallocate the track object and arrays within.                                                          |
+| alloc_track    | Allocate the track object.                                                                              |
[…]

[…]
