+++ docs/source/conf.py
@@ -88,10 +88,16 @@ html_theme = 'sphinx_rtd_theme'
 # html_theme = "nature"
 
 
+html_theme_options = {"display_version": True, "logo_only" : True}
+html_last_updated_fmt = "%Y %b %d at %H:%M:%S UTC"
+html_show_sourcelink = False
+html_logo = "images/metisse_sidebar.png"
+
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+
 
 html_css_files = [
     'custom.css',
