+++ docs/build/html/_sources/index.rst.txt
@@ -3,28 +3,38 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to METISSE's documentation!
-====================================
 
 
-**METISSE** is an acronym for **MET**\ hod of **I**\ nterpolation for **S**\ ingle **S**\ tar **E**\ volution. It is an open-source stellar evolution tool designed to provide a fast and flexible alternative to traditional rapid stellar evolution prescriptions.
+.. raw:: html
+
+   <h1 style="text-align:center;">Welcome to METISSE's documentation!</h1>
+
+.. Welcome to METISSE's documentation!
+.. ====================================
+
+.. figure:: images/metisse_logo_rect.png
+   :width: 60%
+   :align: center
+   :class: largepadding
+
+**METISSE** stands for **MET**\ hod of **I**\ nterpolation for **S**\ ingle **S**\ tar **E**\ volution. It is an open-source stellar evolution tool designed to provide a fast and flexible alternative to traditional rapid stellar evolution prescriptions.
 
 At its core, METISSE interpolates between pre-computed stellar evolution tracks to determine stellar parameters such as luminosity, radius, and core mass at any evolutionary stage. This approach allows users to incorporate stellar models directly into large-scale stellar and binary population synthesis simulations.
 
 
-What METISSE is for
--------------------
+What METISSE is for ?
+--------------------- 
 
 METISSE has been specifically designed to be integrated with binary evolution and population synthesis frameworks that currently rely on the Fortran 77 based SSE `(Hurley et al. 2000) <https://ui.adsabs.harvard.edu/abs/2000MNRAS.315..543H/abstract>`_ codes. METISSE enables such frameworks to:
 
-- Incorporate the latest stellar models from codes like MESA or the BonnCode,
+- Incorporate the latest stellar models,
 - Vary stellar physics assumptions systematically,
-- Explore how uncertainties in stellar evolution impact populations of single and binary stars.
+- Explore how uncertainties in stellar evolution impact binary populations.
 
 Although METISSE can be run as a standalone program to evolve individual stars or simple stellar populations, its main strength lies in serving as an stellar engine within population synthesis environments.
 
-What you can do with METISSE
-----------------------------
+What you can do with METISSE ?
+------------------------------ 
 
 - Run METISSE independently to evolve single stellar systems from pre-computed tracks
 - Use METISSE with supported binary evolution codes to evolve stars in binaries 
@@ -49,11 +59,11 @@ What you can do with METISSE
    <h3>Learn more</h3>
 
    <ul>
-     <li><a href="installation.html"><b>Installation and setup</b> — how to download METISSE</a></li>
-     <li><a href="usage_standalone.html"><b>Single star evolution</b> with METISSE</a></li>
-     <li><a href="usage_other.html"><b>Binary evolution and population synthesis</b> with METISSE as the stellar engine</a></li>
-     <li><a href="using_custom_input_tracks.html"><b>Input stellar tracks</b> — how to use your own models</a></li>
-     <li><a href="structure.html"><b>Structure and Integration</b> — how METISSE works under the hood</a></li>
+     <li><a href="installation.html"><b>Installation</b> — how to download METISSE</a></li>
+     <li><a href="usage_standalone.html">Use METISSE to <b>evolve single stars </b></a></li>
+     <li><a href="usage_other.html">Use METISSE as the stellar engine to <b>evolve binaries </b></a></li>
+     <li><a href="using_custom_input_tracks.html"><b>Use custom tracks</b> — how to use your own tracks with METISSE</a></li>
[…]
