### 2025-10-30 3c16e47
fixed main input to match docs
