+++ example_files/BoOST/MW_BoOST_metallicity.in
@@ -1,0 +1,20 @@
+! Make sure to ADD a new line at the end (after / symbol)
+
+&metallicity_controls
+
+    ! Location of the folder containing input files for a given metallicity
+    eep_tracks_dir = './MW-fixed-core-mass'
+
+    ! Metallicity of input tracks
+    ! It is only used to cross-check 
+    ! against input Z to avoid mistakes
+
+    
+    z_files = 0.008
+
+    ! Details about the file structure (see format_defaults.dat)
+    ! Empty string will raise an error
+
+    format_file = './format_stellar_model_grid'
+
+/
