### 2025-10-30 536285b
added dynamic creation of output folder
