+++ src/METISSE_zcnsts.f90
@@ -394,6 +394,17 @@ subroutine METISSE_zcnsts(z, zpars, ierr)
     
     ! for main, commons are assigned within the METISSE_main
     if (front_end > main) call assign_commons()
+
+    ! create output folders if they do not exist already 
+    ! and output files are to be written
+    if (write_output_to_file .and. front_end <= main) call check_folder('output',ierr)
+    if (write_eep_file) call check_folder('output_eep',ierr)
+
+    if (ierr/=0) then
+        print*, 'METISSE error: cannot create output folder.'
+        return
+    endif
+
     if (debug) print*, 'Finished in zcsnts'
 
         
