### 2025-10-30 3b1afa5
METISSE_dir set dynamically within code
