### 2025-11-01 72f1e93
added demo notebook and modified main output
