+++ docs/source/demo.md
@@ -1,0 +1,20 @@
+# Examples: Running METISSE
+
+This page demonstrates how to run METISSE from a Jupyter notebook, generate Hertzsprung–Russell diagrams, maximum stellar radii, and remnant mass plots using different stellar track libraries (MESA, MIST, BoOST).
+
+## Notebook Demo
+
+The demo notebook is included below. It contains executable Python code, plots, and descriptions.
+## Notes
+
+* Make sure the paths in the notebook for track libraries are correct relative to the documentation build directory.
+* All plots are generated automatically during notebook execution.
+* The notebook uses `metisse_wrapper` to interface with the METISSE Fortran code.
+
+
+```{toctree}
+:maxdepth: 2
+:caption: Contents:
+
+notebook/demo.ipynb
+
