### 2025-11-01 36bf34c
cleaned up demo
