### 2025-11-02 8c84acd
split docfiles for bse and cosmic
