+++ docs/source/usage_other.md
@@ -1,245 +1,12 @@
 # Using METISSE to evolve binaries
 
-METISSE, by itself, is not capable of evolving stars in binary systems. However, it has been specifically designed to be integrated with binary evolution codes that can do so, enabling them to make use of modern stellar data for binary evolution computations.
+**METISSE**, by itself, is not capable of evolving stars in binary systems.  
+However, it has been specifically designed to be integrated with binary evolution codes that perform such calculations, allowing them to make use of modern stellar models in binary evolution studies.
 
-It can be seamlessly integrated with codes that currently use SSE fitting formulae [(Hurley et al. 2000)](https://ui.adsabs.harvard.edu/abs/2000MNRAS.315..543H/abstract) for single star evolution computations.
-It has already been added to the binary evolution codes BSE [(Hurley et al. 2002)](https://ui.adsabs.harvard.edu/abs/2002MNRAS.329..897H/abstract) and COSMIC [(Breivik et al. 2020)](https://ui.adsabs.harvard.edu/abs/2020ApJ...898...71B/abstract).
+In particular, METISSE can be seamlessly integrated with codes that currently use the **SSE fitting formulae** [(Hurley et al. 2000)](https://ui.adsabs.harvard.edu/abs/2000MNRAS.315..543H/abstract) for single-star evolution computations.
 
+Currently, METISSE is available as a module within the following codes:
 
-
-## COSMIC 
-
-
-1. To use METISSE with COSMIC one needs the GitHub repository for the latest version of COSMIC.
-
-If you don't have COSMIC installed already you can fork it from the [GitHub repository](https://github.com/COSMIC-PopSynth/COSMIC/tree/develop) and clone it on your machine.
-
-
-``` console
-
-$ git clone https://github.com/COSMIC-PopSynth/COSMIC
-
-```
-
-Follow the COSMIC's documentation to [load the cosmic environment](https://cosmic-popsynth.github.io/COSMIC/pages/install.html)
-
-Next, cd into the COSMIC's directory and compile the library: 
-
-
-``` console
-
-$ pip install .
-
-```
-
-
-Try evolving an example binary following the instructions on [this page](https://cosmic-popsynth.github.io/COSMIC/pages/examples.html). 
-
-
-2. Next checkout the METISSE-Integrate branch of COSMIC. In the COSMIC's directory, do:
-
-
-``` console
-
-$ git checkout METISSE-integrate
-
-```
-
-Get the METISSE submodule.
-
-``` console
-
-$ git submodule update --init --recursive
-
-```
-
-
-Re-compile the library
-
-
-``` console
-
-$ pip install .
-
-```
-
-3. METISSE is now ready to use with COSMIC. 
-
-The usual input namelist `METISSE_input_controls` is not read when using COSMIC. Instead a Python dictionary, `SSEDict` is used to set the `stellar_engine` and provide the location of the folders containing {term}`Metallicity File` for hydrogen and helium stars through variables `path_to_tracks` and `path_to_he_tracks`.
-
-``` ipython
-SSEDict = {
-          'stellar_engine': 'metisse', 
-          'path_to_tracks': '/Users/poojan/Downloads/sample_tracks_solarZ/Hydrogen/', 
-          'path_to_he_tracks': '/Users/poojan/Downloads/sample_tracks_solarZ/Helium/' 
-          }
-```
-
-:::{Important}
-
[…]
