### 2025-11-02 94b6d0e
bsedict update for cosmic doc
