+++ docs/source/cosmic.md
@@ -1,147 +1,147 @@
 
 # METISSE with COSMIC 
 
-**COSMIC** [(Breivik et al. 2020)](https://ui.adsabs.harvard.edu/abs/2020ApJ...898...71B/abstract) is a binary population synthesis code designed to model the evolution of stellar systems across a wide range of astrophysical environments. It builds upon the rapid binary evolution algorithms of BSE, extending them with population synthesis capabilities and easy to use Python framework. 
-METISSE can be used as the stellar evolution module inside **COSMIC**.
+COSMIC [(Breivik et al. 2020)](https://ui.adsabs.harvard.edu/abs/2020ApJ...898...71B/abstract) is a binary population synthesis code designed to model the evolution of stellar systems across a wide range of astrophysical environments. It builds upon the rapid binary evolution algorithms of BSE, extending them with population synthesis capabilities and easy to use Python framework. 
+METISSE can be used as the stellar evolution module inside COSMIC.
 
-This section provides a short, step-by-step guide to integrating METISSE with COSMIC.
+This section provides a short, step-by-step guide to using METISSE with COSMIC.
 For more details please refer to COSMIC's [official documentation](https://cosmic-popsynth.github.io/COSMIC/index.html). 
 
 1. Follow COSMIC’s official [installation guide](https://cosmic-popsynth.github.io/COSMIC/pages/install.html) to load the COSMIC environment and activate it in your terminal.
 
-If you don’t already have COSMIC installed, download it from [GitHub](https://github.com/COSMIC-PopSynth/COSMIC/tree/develop).
+    If you don’t already have COSMIC installed, download it from [GitHub](https://github.com/COSMIC-PopSynth/COSMIC/tree/develop).
 
 
-``` console
+    ``` console
 
-$ git clone https://github.com/COSMIC-PopSynth/COSMIC
+    $ git clone https://github.com/COSMIC-PopSynth/COSMIC
 
-```
+    ```
 
-Next, change into the COSMIC directory and compile the library: 
-
-
-``` console
-
-$ pip install .
-
-```
+    Next, change into the COSMIC directory and compile the library: 
 
 
-Try evolving an example binary following the instructions on [this page](https://cosmic-popsynth.github.io/COSMIC/pages/examples.html). 
+    ``` console
+
+    $ pip install .
+
+    ```
+
+
+    Try evolving an example binary following the instructions on [this page](https://cosmic-popsynth.github.io/COSMIC/pages/examples.html). 
 
 
 2. Next checkout the METISSE-Integrate branch of COSMIC. In the COSMIC's directory, do:
 
 
-``` console
+    ``` console
 
-$ git checkout METISSE-integrate
+    $ git checkout METISSE-integrate
 
-```
+    ```
 
-Get the METISSE submodule.
+    Get the METISSE submodule.
 
-``` console
+    ``` console
 
-$ git submodule update --init --recursive
+    $ git submodule update --init --recursive
 
-```
-
-
-Re-compile the library
+    ```
 
 
-``` console
+    Re-compile the library
 
-$ pip install .
 
-```
+    ``` console
 
-METISSE is now ready to use with COSMIC.
[…]
