+++ docs/source/bse.md
@@ -1,69 +1,71 @@
 # METISSE with BSE 
 
-BSE[(Hurley et al. 2002)](https://ui.adsabs.harvard.edu/abs/2002MNRAS.329..897H/abstract) is a rapid binary evolution code that models key binary interaction processes such as mass transfer, common-envelope evolution, and supernovae. The algorithms implemented in BSE have served as the foundation for many subsequent rapid and detailed binary evolution codes.
+BSE [(Hurley et al. 2002)](https://ui.adsabs.harvard.edu/abs/2002MNRAS.329..897H/abstract) is a rapid binary evolution code that models key binary interaction processes such as mass transfer, common-envelope evolution, and supernovae. The algorithms implemented in BSE have served as the foundation for many subsequent rapid and detailed binary evolution codes.
  
 Folow these steps to use METISSE with BSE:
 
 1. Get the METISSE-enabled version of BSE through its [GitHub repository](https://github.com/poojanagrawal/BSE-METISSE).
 
 
-``` console
+    ``` console
 
-$ git clone https://github.com/poojanagrawal/BSE-METISSE
-$ git submodule update --init --recursive
+    $ git clone https://github.com/poojanagrawal/BSE-METISSE
+    $ git submodule update --init --recursive
 
-```
- Alternatively, you can clone BSE and METISSE together: 
+    ```
+    Alternatively, you can clone BSE and METISSE together: 
 
-``` console
+    ``` console
 
-$ git clone --recurse-submodules https://github.com/poojanagrawal/BSE-METISSE.git
+    $ git clone --recurse-submodules https://github.com/poojanagrawal/BSE-METISSE.git
 
-```
+    ```
 
 2. Compile the code using:
 
-``` console
+    ``` console
 
-$ ./mk
+    $ ./mk
 
-```
+    ```
 
 3. Let us compute the evolution of a single binary using BSE-METISSE, say a 2.5 M<sub>$_\odot$</sub> star with a 1 M<sub>$_\odot$</sub> companion, with an initial orbital period of 100 days in a circular orbit for the evolution time of 13.7 Gyr.
 
-When using BSE, METISSE-specific inputs are read through the `METISSE_input_controls` inlist in the *evolve_metisse.in* file. (See [](input_files.md#metisse-input-controls) for a complete list of input options). 
+    When using BSE, METISSE-specific inputs are read through the `METISSE_input_controls` inlist in the *evolve_metisse.in* file. (See [](input_files.md#metisse-input-controls) for a complete list of input options). 
 
-**Important:** When using BSE, the file *evolve_metisse.in* should be located in the same directory as the bse executable. 
+    ::: {Important}
+    When using BSE, the file *evolve_metisse.in* should be located in the same directory as the bse executable. 
+    :::
 
-We supply the location of the folders containing {term}`Metallicity File` for hydrogen and helium stars through `METALLICITY_DIR` and `METALLICITY_DIR_HE`. (Make sure to provide relevant paths based on the location of these folders **on your machine** ). 
+    We supply the location of the folders containing {term}`Metallicity File` for hydrogen and helium stars through `METALLICITY_DIR` and `METALLICITY_DIR_HE`. (Make sure to provide relevant paths based on the location of these folders **on your machine** ). 
 
-``` fortran
[…]

[…]
