### 2025-11-02 e6f60fb
Update .readthedocs.yaml
