+++ .readthedocs.yaml
@@ -13,7 +13,15 @@ build:
     # nodejs: "20"
     # rust: "1.70"
     # golang: "1.20"
-
+  jobs:
+    post_create_environment:
+      - echo "Environment created"
+    post_install:
+      - echo "Installing docs dependencies"
+    
+conda:
+  environment: environment.yml
+  
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
   configuration: docs/source/conf.py
