### 2025-11-02 ae095ef
Create environment.yml
