### 2025-11-02 3d794b4
Update .readthedocs.yaml
