### 2025-11-02 681ce43
Update .readthedocs.yaml
