+++ .readthedocs.yaml
@@ -7,8 +7,8 @@ version: 2
 # Set the OS, Python version and other tools you might need
 build:
   os: ubuntu-22.04
-  tools:
-    python: "3.12"
+  # tools:
+  #   python: "3.12"
     # You can also specify other tool versions:
     # nodejs: "20"
     # rust: "1.70"
