### 2025-11-02 bc14741
Update .readthedocs.yaml
