+++ .readthedocs.yaml
@@ -13,11 +13,11 @@ build:
     # nodejs: "20"
     # rust: "1.70"
     # golang: "1.20"
-  jobs:
-    post_create_environment:
-      - echo "Environment created"
-    post_install:
-      - echo "Installing docs dependencies"
+  commands:
+    - echo "Building docs with sphinx-build"
+    - python -m pip install --upgrade pip setuptools wheel
+    - python -m pip install -e . || true   # optional: install package if you want imports to work
+    - sphinx-build -b html docs/source _build/html
     
 conda:
   environment: environment.yml
