### 2025-11-02 e9c4b6b
Update .readthedocs.yaml
