### 2025-11-02 66877ca
Update .readthedocs.yaml
