+++ .readthedocs.yaml
@@ -14,6 +14,10 @@ build:
     # rust: "1.70"
     # golang: "1.20"
   commands:
+    - echo "===== Build environment info ====="
+    - python --version
+    - echo $PATH
+    - which sphinx-build || true
     - make -C docs html
     
 conda:
