### 2025-11-02 73c02b2
Update .readthedocs.yaml
