+++ .readthedocs.yaml
@@ -7,15 +7,14 @@ version: 2
 # Set the OS, Python version and other tools you might need
 build:
   os: ubuntu-22.04
-  # tools:
-  #   python: "3.12"
+  tools:
+    python: "3.12"
     # You can also specify other tool versions:
     # nodejs: "20"
     # rust: "1.70"
     # golang: "1.20"
   commands:
     - echo "===== Build environment info ====="
-    - python --version
     - echo $PATH
     - which sphinx-build || true
     - make -C docs html
