### 2025-11-02 3867fd7
Update .readthedocs.yaml
