+++ .readthedocs.yaml
@@ -7,17 +7,16 @@ version: 2
 # Set the OS, Python version and other tools you might need
 build:
   os: ubuntu-22.04
-  tools:
-    python: "3.12"
+  # tools:
+  #   python: "3.12"
     # You can also specify other tool versions:
     # nodejs: "20"
     # rust: "1.70"
     # golang: "1.20"
   commands:
-    - echo "===== Build environment info ====="
-    - echo $PATH
-    - which sphinx-build || true
-    - make -C docs html
+  - source /home/docs/checkouts/readthedocs.org/user_builds/metisse/conda/latest/bin/activate
+  - sphinx-build -b html docs/source docs/_build/html
+
     
 conda:
   environment: environment.yml
