### 2025-11-02 b78e6ac
Update .readthedocs.yaml
