### 2025-11-02 38d73dc
Update environment.yml
