+++ environment.yml
@@ -2,21 +2,31 @@ name: docs-env
 channels:
   - conda-forge
 dependencies:
+  # Python version
   - python=3.10
+  
+  # Compilers for Fortran extensions
   - gfortran
   - gcc
+  
+  # Sphinx and extensions
   - sphinx>=5,<8
   - nbsphinx>=0.9
   - jinja2<3.1
   - nbconvert
-  - pandocfilters
-  - pandoc<3.0.0
   - myst_parser==3.0.1
   - sphinx_rtd_theme==2.0.0
   - readthedocs-sphinx-search==0.3.2
   - sphinx_copybutton
   - lxml_html_clean
   - ipykernel
+  
+  # Pandoc for notebook conversion
+  - pandoc<3.0.0
+  - pandocfilters
+  
+  # Scientific stack
   - numpy
 
-
+  # pip included for any pip-only packages
+  - pip
