### 2025-11-02 6a05567
Update .readthedocs.yaml
