+++ .readthedocs.yaml
@@ -14,7 +14,7 @@ build:
     # rust: "1.70"
     # golang: "1.20"
   commands:
-  - source $HOME/conda/latest/bin/activate
+  - . $HOME/conda/latest/bin/activate 
   - sphinx-build -b html docs/source docs/_build/html
 
     
