### 2025-11-02 6394f54
Update .readthedocs.yaml
