+++ .readthedocs.yaml
@@ -14,7 +14,6 @@ build:
     # rust: "1.70"
     # golang: "1.20"
   commands:
-  - . $HOME/conda/latest/bin/activate 
   - sphinx-build -b html docs/source docs/_build/html
 
     
