### 2025-11-02 2b39c17
Update .readthedocs.yaml
