### 2025-11-02 300efac
Update .readthedocs.yaml
