+++ docs/requirements.txt
@@ -8,4 +8,3 @@ lxml_html_clean
 pandoc<3.0.0
 ipykernel 
 numpy
-gfortran
