+++ environment.yml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
 dependencies:
   # Python version
-  - python=3.10
+  - python=3.12.10
   
   # Compilers for Fortran extensions
   - gfortran
