### 2025-11-02 23bb603
Update .readthedocs.yaml
