### 2025-11-02 60b0634
Update .readthedocs.yaml
