### 2025-11-02 16fb53e
Update .readthedocs.yaml
