+++ .readthedocs.yaml
@@ -7,12 +7,9 @@ version: 2
 # Set the OS, Python version and other tools you might need
 build:
   os: ubuntu-22.04
+  os: "ubuntu-24.04"
   tools:
-    python: "3.12"
-    # You can also specify other tool versions:
-    # nodejs: "20"
-    # rust: "1.70"
-    # golang: "1.20"
+    python: "miniconda3-3.12-24.9"
   jobs:
     post_install:
       - echo "===== Build environment info ====="
