### 2025-11-02 3709858
Update .readthedocs.yaml
