+++ .readthedocs.yaml
@@ -6,7 +6,6 @@ version: 2
 
 # Set the OS, Python version and other tools you might need
 build:
-  os: ubuntu-22.04
   os: "ubuntu-24.04"
   tools:
     python: "miniconda3-3.12-24.9"
