### 2025-11-02 9a7389e
Update .readthedocs.yaml
