### 2025-11-02 41b518c
Update environment.yml
