### 2025-11-02 a7904a1
Update .readthedocs.yaml
