### 2025-11-02 27419a2
Update environment.yml
