### 2025-11-03 ac59748
Update .readthedocs.yaml
