+++ .readthedocs.yaml
@@ -9,12 +9,12 @@ build:
   os: "ubuntu-24.04"
   tools:
     python: "miniconda3-3.12-24.9"
-  jobs:
-    post_install:
-      - make -C docs html
+#   jobs:
+#     post_install:
+#       - make -C docs html
     
-conda:
-  environment: environment.yml
+# conda:
+#   environment: environment.yml
   
 # Build documentation in the "docs/" directory with Sphinx
 sphinx:
@@ -32,6 +32,6 @@ sphinx:
 # Optional but recommended, declare the Python requirements required
 # to build your documentation
 # See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
-# python:
-#   install:
-#     - requirements: docs/requirements.txt
+python:
+  install:
+    - requirements: docs/requirements.txt
