+++ docs/requirements.txt
@@ -7,4 +7,3 @@ nbsphinx>=0.8.6
 lxml_html_clean
 pandoc<3.0.0
 ipykernel 
-numpy
