### 2025-11-03 21c927d
Update .readthedocs.yaml
