+++ .readthedocs.yaml
@@ -8,7 +8,8 @@ version: 2
 build:
   os: "ubuntu-24.04"
   tools:
-    python: "miniconda3-3.12-24.9"
+    python: "3.12"
+    # python: "miniconda3-3.12-24.9"
 #   jobs:
 #     post_install:
 #       - make -C docs html
