### 2025-11-03 4502328
updated paper draft
