+++ paper.bib
@@ -39,6 +39,36 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Rodriguez:2021,
+       author = {{Rodriguez}, Carl L. and {Weatherford}, Newlin C. and {Coughlin}, Scott C. and {Amaro-Seoane}, Pau and {Breivik}, Katelyn and {Chatterjee}, Sourav and {Fragione}, Giacomo and {K{\i}ro{\u{g}}lu}, Fulya and {Kremer}, Kyle and {Rui}, Nicholas Z. and {Ye}, Claire S. and {Zevin}, Michael and {Rasio}, Frederic A.},
+        title = "{Modeling Dense Star Clusters in the Milky Way and beyond with the Cluster Monte Carlo Code}",
+      journal = {\apjs},
+     keywords = {1883, 1971, 678, 656, 1083, 1596, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2022,
+        month = feb,
+       volume = {258},
+       number = {2},
+          eid = {22},
+        pages = {22},
+          doi = {10.3847/1538-4365/ac2edf},
+archivePrefix = {arXiv},
+       eprint = {2106.02643},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022ApJS..258...22R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@book{Aarseth:2003, 
+      place={Cambridge}, 
+      series={Cambridge Monographs on Mathematical Physics}, 
+      title={Gravitational N-Body Simulations: Tools and Algorithms}, 
+      DOI={10.1017/CBO9780511535246}, 
+      publisher={Cambridge University Press}, 
+      author={Aarseth, Sverre J.}, 
+      year={2003}, 
+      collection={Cambridge Monographs on Mathematical Physics}
+}
+
 @ARTICLE{Agrawal:2020,
        author = {{Agrawal}, Poojan and {Hurley}, Jarrod and {Stevenson}, Simon and
          {Sz{\'e}csi}, Dorottya and {Flynn}, Chris},
@@ -223,3 +253,59 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Fragos:2022,
+       author = {{Fragos}, Tassos and {Andrews}, Jeff J. and {Bavera}, Simone S. and {Berry}, Christopher P.~L. and {Coughlin}, Scott and {Dotter}, Aaron and {Giri}, Prabin and {Kalogera}, Vicky and {Katsaggelos}, Aggelos and {Kovlakas}, Konstantinos and {Lalvani}, Shamal and {Misra}, Devina and {Srivastava}, Philipp M. and {Qin}, Ying and {Rocha}, Kyle A. and {Rom{\'a}n-Garza}, Jaime and {Serra}, Juan Gabriel and {Stahle}, Petter and {Sun}, Meng and {Teng}, Xu and {Trajcevski}, Goce and {Tran}, Nam Hai and {Xing}, Zepei and {Zapartas}, Emmanouil and {Zevin}, Michael},
+        title = "{POSYDON: A General-purpose Population Synthesis Code with Detailed Binary-evolution Simulations}",
+      journal = {\apjs},
[…]
