+++ paper.md
@@ -46,32 +46,32 @@ affiliations:
    index: 6
 
 
-date: 16 Feb 2025
+date: 3 Nov 2025
 bibliography: paper.bib
 
 ---
 
 # Summary
 
-METISSE is an open-source stellar evolution tool specifically designed to be integrated with binary evolution and population synthesis codes. Unlike traditional rapid stellar evolution schemes based on static fitting formulae (e.g. SSE; [@Hurley:2000]), METISSE interpolates between pre-computed stellar models to quickly derive stellar parameters for population synthesis codes. This approach makes it straightforward to incorporate the latest stellar physics and explore the effects of different input models. METISSE can be seamlessly integrated with any population synthesis code that currently uses popular Fortran code SSE [@Hurley:2000] to calculate stellar parameters. While METISSE can also be used in standalone mode to evolve single stellar populations, its primary purpose is to act as an interpolation engine within existing population synthesis codes for modelling binary stars and star clusters. Written in Modern Fortran, METISSE is both fast and robust, making it well-suited for large-scale population studies that require realistic stellar evolution input. The full documentation for METISSE is available at []https://metisse.readthedocs.io.
+METISSE is an open-source stellar evolution tool specifically designed to be integrated with binary evolution and population synthesis codes. Unlike traditional rapid stellar evolution schemes based on the fitting formulae (e.g. SSE; [@Hurley:2000]), METISSE interpolates between pre-computed stellar models to quickly derive stellar parameters for population synthesis codes. This approach makes it straightforward to incorporate the latest stellar models and explore the effects of different input physics. METISSE can be easily integrated with any population synthesis code that currently uses popular Fortran code SSE to calculate stellar parameters. While METISSE can also be used in standalone mode to evolve single stellar populations, its primary purpose is to act as an interpolation engine within existing population synthesis codes for modelling binary stars and star clusters. Written in Modern Fortran, METISSE is both fast and robust, making it well-suited for large-scale population studies that require realistic stellar evolution input. The full documentation for METISSE is available at [](https://metisse.readthedocs.io).
 
 # Statement of need
 
-Stars, especially those with masses greater than eight solar masses (massive stars), play a pivotal role in shaping stellar populations. The best way of computing stellar evolution involves solving equations of stellar structure and evolution through detailed stellar evolution codes such as MESA. However, the inherent uncertainties in stellar evolution cause stellar codes to adopt different physical inputs, leading to significant differences in the predictions for the evolution of stars and stellar populations [@Agrawal:2022a]. Moreover, computational requirements and robustness issues render these codes impractical for direct use in large population synthesis simulations. 
+Stars, especially those with masses greater than eight solar masses (massive stars), play a pivotal role in shaping stellar populations. The best way of computing stellar evolution involves solving equations of stellar structure and evolution through detailed stellar evolution codes such as MESA [@Jermyn:2023]. However, the inherent uncertainties in stellar evolution cause stellar codes to adopt different physical inputs, leading to significant differences in the predictions for the evolution of stars and stellar populations [@Agrawal:2022a]. Moreover, computational requirements and robustness issues render these codes impractical for direct use in large population synthesis simulations. 
 
[…]
