+++ paper.md
@@ -53,7 +53,7 @@ bibliography: paper.bib
 
 # Summary
 
-METISSE is an open-source stellar evolution tool specifically designed to be integrated with binary evolution and population synthesis codes. Unlike traditional rapid stellar evolution schemes based on the fitting formulae (e.g. SSE; [@Hurley:2000]), METISSE interpolates between pre-computed stellar models to quickly derive stellar parameters for population synthesis codes. This approach makes it straightforward to incorporate the latest stellar models and explore the effects of different input physics. METISSE can be easily integrated with any population synthesis code that currently uses popular Fortran code SSE to calculate stellar parameters. While METISSE can also be used in standalone mode to evolve single stellar populations, its primary purpose is to act as an interpolation engine within existing population synthesis codes for modelling binary stars and star clusters. Written in Modern Fortran, METISSE is both fast and robust, making it well-suited for large-scale population studies that require realistic stellar evolution input. The full documentation for METISSE is available at [](https://metisse.readthedocs.io).
+METISSE is an open-source stellar evolution tool specifically designed to be integrated with binary evolution and population synthesis codes. Unlike traditional rapid stellar evolution schemes based on the fitting formulae (e.g. SSE; [@Hurley:2000]), METISSE interpolates between pre-computed stellar models to quickly derive stellar parameters for population synthesis codes. This approach makes it straightforward to incorporate the latest stellar models and explore the effects of different input physics. METISSE can be easily integrated with any population synthesis code that currently uses popular Fortran code SSE to calculate stellar parameters. While METISSE can also be used in standalone mode to evolve single stellar populations, its primary purpose is to act as an interpolation engine within existing population synthesis codes for modelling binary stars and star clusters. Written in Modern Fortran, METISSE is both fast and robust, making it well-suited for large-scale population studies that require realistic stellar evolution input. The full documentation for METISSE is available at [metisse.readthedocs.io](https://metisse.readthedocs.io).
 
 # Statement of need
 
@@ -63,7 +63,7 @@ Rapid stellar-evolution codes such as SSE [@Hurley:2000]that rely on fitting for
 
 In recent years, interpolation-based rapid stellar-evolution codes such as TRILEGAL [@Girardi:2005], ComBinE [@Kruckow:2018], SEVN [@Iorio:2022], and MINT [@Rees:2025], along with machine-learning–based frameworks like POSYDON [@Fragos:2022], have become increasingly common as alternatives to SSE. However, these tools can not be easily incorporated into existing frameworks: binary population codes such as BSE [@Hurley:2002] and COSMIC [@Breivik:2020], or star cluster codes such as CMC [Rodriguez:2021] and NBODY6 [Aarseth:2003], that continue to rely on standard SSE routines.
 
-METISSE provides performance and capabilities comparable to other interpolation-based codes but is implemented in Modern Fortran with full Fortran 77 compatibility. It includes SSE-style subroutines for straightforward integration into existing binary and population-synthesis codes. METISSE comes with an example set of [MESA stellar models](https://zenodo.org/records/17513335) and can also use other published grids, such as MIST [Choi:2016] and BoOST [@Szecsi:2020]. Its interpolation framework allows stellar model grids to be easily swapped and explore how stellar parameters affect population outcomes. METISSE thus offers a flexible and efficient framework for stellar-evolution calculations in population synthesis, offering higher physical accuracy than analytic prescriptions without the heavy computational cost of full stellar-structure calculations. 
[…]
