+++ paper.md
@@ -61,7 +61,7 @@ Stars, especially those with masses greater than eight solar masses (massive sta
 
 Rapid stellar-evolution codes such as SSE [@Hurley:2000]that rely on fitting formulae manually calibrated to specific stellar models, have long provided a fast and efficient way to calculate stellar population properties. However, because the formulas must be recalculated manually for each new set of stellar models, they cannot be easily updated to incorporate advances in stellar evolution, limiting their flexibility and applicability to modern studies.
 
-In recent years, interpolation-based rapid stellar-evolution codes such as TRILEGAL [@Girardi:2005], ComBinE [@Kruckow:2018], SEVN [@Iorio:2022], and MINT [@Rees:2025], along with machine-learning–based frameworks like POSYDON [@Fragos:2022], have become increasingly common as alternatives to SSE. However, these tools can not be easily incorporated into existing frameworks: binary population codes such as BSE [@Hurley:2002] and COSMIC [@Breivik:2020], or star cluster codes such as CMC [Rodriguez:2021] and NBODY6 [Aarseth:2003], that continue to rely on standard SSE routines.
+In recent years, interpolation-based rapid stellar-evolution codes such as TRILEGAL [@Girardi:2005], ComBinE [@Kruckow:2018], SEVN [@Iorio:2022], and MINT [@Rees:2025], along with machine-learning–based frameworks like POSYDON [@Fragos:2022], have become increasingly common as alternatives to SSE. However, these tools can not be easily incorporated into existing frameworks: binary population codes such as BSE [@Hurley:2002] and COSMIC [@Breivik:2020], or star cluster codes such as CMC [@Rodriguez:2021] and NBODY6 [@Aarseth:2003], that continue to rely on standard SSE routines.
 
 METISSE provides performance and capabilities comparable to other interpolation-based codes but is implemented in Modern Fortran with full Fortran 77 compatibility. It includes SSE-style subroutines for straightforward integration into existing binary and population-synthesis codes. METISSE comes with an example set of [MESA stellar models](https://zenodo.org/records/17513335) and can also use other published grids, such as MIST [@Choi:2016] and BoOST [@Szecsi:2020]. Its interpolation framework allows stellar model grids to be easily swapped and explore how stellar parameters affect population outcomes. METISSE thus offers a flexible and efficient framework for stellar-evolution calculations in population synthesis, offering higher physical accuracy than analytic prescriptions without the heavy computational cost of full stellar-structure calculations. 
 
