### 2025-11-14 TeamMETISSE/METISSE#90 @TomWagg
`output_binary` directory not defined
I got this error when running the BSE example

```
Fortran runtime error: Cannot open file 'output_binary/00250M_00100M.dat': No such file or directory

Error termination. Backtrace:
#0  0x14712335f171 in ???
#1  0x14712335fd19 in ???
#2  0x147123360521 in ???
#3  0x147123564288 in ???
#4  0x14712356458c in ???
#5  0x402191 in bse
        at ../src/bse.f:200
#6  0x40173c in main
        at ../src/bse.f:242
```

This is just because the `output_binary` folder isn't present in the GitHub repo. (You can add a README.md file to an empty folder to make it appear on GitHub and describe what will be in it).

Should be a quick fix :)

(Part of https://github.com/openjournals/joss-reviews/issues/8817)
