### 2025-11-14 TeamMETISSE/METISSE#91 @TomWagg
OPTIONAL: Include a comparison example between SSE and METISSE when running COSMIC?
Here's a notebook I made to play around with this, feel free to use any/all of the code if you want to.

[metisse_cosmic_example.ipynb](https://github.com/user-attachments/files/23553726/metisse_cosmic_example.ipynb)

(Part of https://github.com/openjournals/joss-reviews/issues/8817)
