### 2025-11-17 4fa7b20
added return when age exceeds max_age
