### 2025-11-17 5e8d494
renamed some vars to more sensible names
