+++ src/interp_support.f90
@@ -6,7 +6,7 @@ module interp_support
     integer, parameter:: linear = 1
     integer, parameter:: Steffen1990 = 2
     !integer, parameter:: extrapolation = 3
-    logical:: debug_mass
+    logical:: debug_mass, debug_age
 
     contains
 
@@ -617,22 +617,20 @@ module interp_support
         integer:: j, k, mlo, mhi, pass, n_pass
         
         real(dp), allocatable:: new_line(:,:)
-        integer, allocatable:: min_eeps(:)
+        integer, allocatable:: nbr_eeps(:)
 
-        logical:: debug
+        debug_age = .false.
+!         if (t% is_he_track) debug_age = .true.
+!        if (t% pars% phase >= 4 .and. (present(icolumn).eqv..false.)) debug_age = .true.
 
-        debug = .false.
-!         if (t% is_he_track) debug = .true.
-!        if (t% pars% phase >= 4 .and. (present(icolumn).eqv..false.)) debug = .true.
-
-        if (debug) print*,"in interpolate age",t% pars% phase
+        if (debug_age) print*,"in interpolate age",t% pars% phase
         dx = 0d0; alfa = 0d0; beta = 0d0; x = 0d0; y = 0d0
         jstart = 1
         jend = t% ncol
         if (present(icolumn)) then
             jstart = icolumn
             jend = icolumn
-            if (debug) print*,"only interpolating in column number",icolumn
+            if (debug_age) print*,"only interpolating in column number",icolumn
         endif
         
         allocate (new_line(t% ncol, 1))
@@ -656,17 +654,17 @@ module interp_support
                 if (t% is_he_track) age_col = i_he_age
             endif
             
-            call find_nearest_eeps(t, min_eeps, age, age_col)
+            call find_neighboring_eeps(t, nbr_eeps, age, age_col)
 
-            mlo = minval(min_eeps)
-            mhi = maxval(min_eeps)
+            mlo = minval(nbr_eeps)
+            mhi = maxval(nbr_eeps)
 
-            if (debug) print*, 'pass', n_pass, pass, age_col, age, t% pars% phase
-            if (debug) print*,"neighbouring_eeps", min_eeps
-!            if (debug) print*,"ages", t% tr(age_col, mlo:mhi)
+            if (debug_age) print*, 'pass', n_pass, pass, age_col, age, t% pars% phase
+            if (debug_age) print*,"neighbouring_eeps", nbr_eeps
+!            if (debug_age) print*,"ages", t% tr(age_col, mlo:mhi)
 
             if (mhi == mlo) then
-                if (debug) print*, "no interp in age needed"
+                if (debug_age) print*, "no interp in age needed"
                 do j = jstart, jend
                     if (pass == 2) then
                         ! check if it is core-related quantity
@@ -679,7 +677,7 @@ module interp_support
                 !linear interpolation
                 alfa = (age-t% tr(age_col, mlo))/(t% tr(age_col, mhi) - t% tr(age_col, mlo))
                 beta = 1d0-alfa
-                if (debug) print*, "doing linear interp in age",alfa, age, t% tr(age_col, mlo), t% tr(age_col, mhi)
+                if (debug_age) print*, "doing linear interp in age",alfa, age, t% tr(age_col, mlo), t% tr(age_col, mhi)
 
                 do j = jstart, jend
                     if (pass == 2) then
@@ -696,11 +694,11 @@ module interp_support
                         return
                     endif
                 end do
-                if (debug) print*, "ending linear interp in age"
+                if (debug_age) print*, "ending linear interp in age"
 
             else
                 ! currently only linear interpolation is used
[…]
