### 2025-11-17 0e083f1
fixed wd calculations
