+++ src/METISSE_hrdiag.f90
@@ -56,6 +56,7 @@
         if (t% post_agb) then
             ! contruct an artficial track until WD cooling phase is reached
             call evolve_after_agb(t)
+            if (t% post_agb .eqv. .false.) has_become_remnant = .true.
         else
             !check if phase/type/kw of the star has changed
 
@@ -80,6 +81,7 @@
             else
                 t% pars% mass = mt
             endif
+            
             IF (check_ge(t% pars% age,t% times(11))) THEN
                 !check if have reached the end of the eep track
                 if (debug)print*,"end of file:aj,tn ",t% pars% age,t% times(11),t% times(max(kw,1))
@@ -90,30 +92,25 @@
                 ! mc_max = MAX(M_ch,0.773* Mcbagb-0.35)
 
                 t% pars% core_mass = t% pars% McCO
-                if ((t% initial_mass.gt.very_low_mass_limit) .and. (t% pars% core_mass<tiny)) then
+                if (t% initial_mass<=very_low_mass_limit) then
+                    t% pars% phase = HeWD
+                elseif (t% pars% core_mass<tiny) then
                     write(UNIT=err_unit,fmt=*)"METISSE error: non-positive core mass",t% pars% core_mass
                     code_error = .true.
                     !assigning an ad-hoc non-zero core mass so the code doesn't break
                     t% pars% core_mass = 0.75*t% pars% McHe
                 endif
+                
                 has_become_remnant = .true.
             
             ELSEIF (check_ge(t% pars% core_mass,t% pars% mass)) THEN
                 !check if envelope has been lost
-    
+
                 if (debug)print*,"envelope lost at",t% pars% age,t% pars% phase,t% pars% mass,t% pars% core_mass
+                call assign_stripped_star_phase(t, HeI_time)
                 
-                if (t% pars% phase == TPAGB) then
-                    ! TPAGB star becomes a CO-WD/ONe WD upon losing envelope
-                    j_bagb = min(t% ntrack, TA_cHeB_EEP)
-                    Mcbagb = t% tr(i_he_core, j_bagb)
-                    has_become_remnant = .true.
-                    !TODO: add a check if it's not a white dwarf
-                else
-                    call assign_stripped_star_phase(t, HeI_time)
-                    if(t% pars% phase == HeWD) then
-                        has_become_remnant = .true.
-                    elseif (use_sse_NHe) then
+                if(t% pars% phase>=He_MS .and. t% pars% phase<=He_GB) then
+                    if (use_sse_NHe) then
                         t% star_type = sse_he_star
                         t% zams_mass = t% pars% mass
 
@@ -131,10 +128,16 @@
                         endif
                         ! zams_mass is assigned in the star
                         call METISSE_star(t% pars% phase, mass,t% pars% mass,tm,tn,tscls,lums,GB,zpars,0.d0,id)
-!                        t% pars% age_old = t% pars% age
+    !                        t% pars% age_old = t% pars% age
 
                         if (debug)print*, 'after env loss', t% pars% phase, t% pars% age,t% MS_time,t% nuc_time,id
                     endif
+                else
+                    j_bagb = min(t% ntrack, TA_cHeB_EEP)
+                    Mcbagb = t% tr(i_he_core, j_bagb)
+                    has_become_remnant = .true.
+                    ! below is so that we accidentally don't recreate postagb phase for wds
+                    t% star_type = remnant
                 endif
             ELSE
                 ! Calculate mass and radius of convective envelope, and envelope gyration radius.
@@ -158,6 +161,7 @@
                 if ((t% pars% core_mass >=mc_max) .or. (abs(mc_max-t% pars% core_mass)<tiny)) then
                     t% pars% core_mass = mc_max
                     has_become_remnant = .true.
[…]
