### 2025-11-17 8f6cb4d
fixed wd calculations
