### 2025-11-17 8833b00
destroy stars below 1e-4 msun
