### 2025-11-18 TeamMETISSE/METISSE#91 @poojanagrawal
Hi Tom, this is a great notebook and I am thinking of adding it with minor changes to the example usage page (with proper acknowledgement).
You happy with me doing it?
