### 2025-11-18 TeamMETISSE/METISSE#91 @TomWagg
Yeah feel free! Though I left it pretty empty of comments sorry haha 😅

I had a lot of fun testing this out, very excited to be using METISSE + COSMIC soon 🎉
