### 2025-11-18 b4f9810
added tom_s notebook
