+++ paper.md
@@ -71,7 +71,7 @@ Multiple ongoing projects use METISSE alongside the binary population synthesis 
 In the era of big-data astronomy, driven by high-quality observational data from both ground-based and space-based telescopes, as well as gravitational wave and multi-messenger detectors, METISSE facilitates the seamless incorporation of updates in stellar evolution into simulations that model stellar populations and their interactions.
 
 # Acknowledgements
-We thank Duncan P. Maclean, Christopher Crow, Runqiu Ye and Steven Rieder for their help with testing METISSE. PA acknowledges support from the European Research Council (ERC) under the Horizon Europe programme (Synergy Grant agreement 101071505: 4D-STAR). While partially funded by the European Union, views and opinions expressed are however those of the author only and do not necessarily reflect those of the European Union or the European Research Council. Neither the European Union nor the granting authority can be held responsible for them. PA, JH and SS acknowledge support from the Australian Research Council Centre of Excellence for Gravitational Wave Discovery (OzGrav), through project numbers CE170100004 and CE230100016. CR acknowledges support from the Alfred P. Sloan Foundation and the David and Lucile Packard Foundation. AK acknowledges financial support
+We thank Duncan Maclean, Christopher Crow, Runqiu Ye and Steven Rieder for their help with testing METISSE. PA acknowledges support from the European Research Council (ERC) under the Horizon Europe programme (Synergy Grant agreement 101071505: 4D-STAR). While partially funded by the European Union, views and opinions expressed are however those of the author only and do not necessarily reflect those of the European Union or the European Research Council. Neither the European Union nor the granting authority can be held responsible for them. PA, JH and SS acknowledge support from the Australian Research Council Centre of Excellence for Gravitational Wave Discovery (OzGrav), through project numbers CE170100004 and CE230100016. CR acknowledges support from the Alfred P. Sloan Foundation and the David and Lucile Packard Foundation. AK acknowledges financial support
 from the Flemish Government under the long-term structural Methusalem
 funding program by means of the project SOUL: Stellar evolution in full
 glory, grant METH/24/012 at KU Leuven. SS is supported by the ARC Discovery Early Career Research Award DE220100241. DSz acknowledges support from the National Science Center (NCN), Poland under grant number OPUS 2021/41/B/ST9/00757.
