+++ docs/source/structure.md
@@ -185,10 +185,8 @@ Update your Makefile or compilation instructions to:
 
 Depending on the code, additional subroutines related to stellar and binary evolution (e.g., evolv2.f and comenv.f) may require minor modifications to work with METISSE. 
 
-<!-- | comenv_lambda.f90   | Get the appropriate ZAMS radius and calculate common envelope lambda (only for binaries).       -->
-
-In case of any doubts, you can refer to how METISSE has been added to COSMIC and BSE. 
-For help, please get in touch through GitHub [discussions](https://github.com/TeamMETISSE/METISSE/discussions).
+If you are unsure about the required changes, you can refer to the integrations of METISSE within COSMIC and BSE as examples. You can find usage examples on the [Using METISSE with binary evolution codes](usage_other.md) page.
+If you need help, feel free to reach out through [GitHub Discussions](https://github.com/TeamMETISSE/METISSE/discussions).
 
 
 <!-- # module details
