+++ docs/source/demo.md
@@ -3,7 +3,7 @@
 This section demonstrates how to use METISSE right from a Jupyter notebook. You will be able to explore stellar evolution, generate Hertzsprung–Russell diagrams, examine maximum stellar radii, and create remnant mass plots using different stellar track libraries, all without leaving Python.
 
 ```{toctree}
+:maxdepth: 1
 
 notebook/demo.ipynb
-
 notebook/metisse_cosmic_example.ipynb
