### 2025-11-19 4aa6056
removed paper files
