+++ paper.md
@@ -1,80 +1,0 @@
----
-title: 'METISSE: METhod of Interpolation for Single Star Evolution'
-tags:
-  - Fortran
-  - astronomy
-  - stellar populations
-  - stellar evolution
-  - stars
-authors:
-  - name: Poojan Agrawal
-    orcid: 0000-0002-1135-984X
-    affiliation: 1 
-  - name: Katelyn Breivik
-    orcid: 0000-0001-5228-6598
-    affiliation: 3 
-  - name: Jarrod Hurley
-    orcid: 0000-0003-2694-0415
-    affiliation: "4,5" # (Multiple affiliations must be quoted)
-  - name: Carl Rodriguez
-    orcid: 0000-0003-4175-8881
-    affiliation: 2 
-  - name: Simon Stevenson
-    orcid: 0000-0002-6100-537X
-    affiliation: "4,5" 
-  - name: Alex Kemp
-    orcid: 0000-0003-2059-5841
-    affiliation: 1
-  - name: Dorottya Szécsi
-    orcid: 0000-0001-6473-7085
-    affiliation: 6
-
-    
-
-affiliations:
- - name: Institute of Astronomy, KU Leuven, Celestijnenlaan 200D, B-3001, Leuven, Belgium
-   index: 1
- - name: Department of Physics and Astronomy, University of North Carolina at Chapel Hill, 120 E. Cameron Avenue, Chapel Hill, NC 27599, USA
-   index: 2
- - name: McWilliams Center for Cosmology, Department of Physics, Carnegie Mellon University, 5000 Forbes Avenue, Pittsburgh, PA 15213, USA
-   index: 3
- - name: Centre for Astrophysics and Supercomputing, Swinburne University of Technology, Hawthorn, VIC 3122, Australia
-   index: 4 
- - name: OzGrav-The ARC Centre of Excellence for Gravitational Wave Discovery, Hawthorn, VIC 3122, Australia
-   index: 5
- - name: Institute of Astronomy, Faculty of Physics, Astronomy and Informatics, Nicolaus Copernicus University, Grudziadzka 5, 87-100, Torun, Poland
-   index: 6
-
-
-date: 3 Nov 2025
-bibliography: paper.bib
-
----
-
-# Summary
-
-METISSE is an open-source stellar evolution tool specifically designed to be integrated with binary evolution and population synthesis codes. Unlike traditional rapid stellar evolution schemes based on the fitting formulae (e.g. SSE; [@Hurley:2000]), METISSE interpolates between pre-computed stellar models to quickly derive stellar parameters for population synthesis codes. This approach makes it straightforward to incorporate the latest stellar models and explore the effects of different input physics. METISSE can be easily integrated with any population synthesis code that currently uses popular Fortran code SSE to calculate stellar parameters. While METISSE can also be used in standalone mode to evolve single stellar populations, its primary purpose is to act as an interpolation engine within existing population synthesis codes for modelling binary stars and star clusters. Written in Modern Fortran, METISSE is both fast and robust, making it well-suited for large-scale population studies that require realistic stellar evolution input. The full documentation for METISSE is available at [metisse.readthedocs.io](https://metisse.readthedocs.io).
-
-# Statement of need
-
-Stars, especially those with masses greater than eight solar masses (massive stars), play a pivotal role in shaping stellar populations. The best way of computing stellar evolution involves solving equations of stellar structure and evolution through detailed stellar evolution codes such as MESA [@Jermyn:2023]. However, the inherent uncertainties in stellar evolution cause stellar codes to adopt different physical inputs, leading to significant differences in the predictions for the evolution of stars and stellar populations [@Agrawal:2022a]. Moreover, computational requirements and robustness issues render these codes impractical for direct use in large population synthesis simulations. 
-
[…]
