+++ paper.bib
@@ -1,311 +1,0 @@
-
-
-
-@article{Hurley:2000,
-       author = {{Hurley}, Jarrod R. and {Pols}, Onno R. and {Tout}, Christopher A.},
-        title = "{Comprehensive analytic formulae for stellar evolution as a function of mass and metallicity}",
-      journal = {Monthly Notices of the Royal Astronomical Society},
-     keywords = {METHODS: ANALYTICAL, STARS: EVOLUTION, STARS: FUNDAMENTAL PARAMETERS, STARS: MASS-LOSS, STARS: POPULATION II, GALAXIES: STELLAR CONTENT, Astrophysics},
-         year = 2000,
-        month = jul,
-       volume = {315},
-       number = {3},
-        pages = {543-569},
-          doi = {10.1046/j.1365-8711.2000.03426.x},
-archivePrefix = {arXiv},
-       eprint = {astro-ph/0001295},
- primaryClass = {astro-ph},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2000MNRAS.315..543H},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-      
- 
-}
-      
-@article{Hurley:2002,
-       author = {{Hurley}, Jarrod R. and {Tout}, Christopher A. and {Pols}, Onno R.},
-        title = "{Evolution of binary stars and the effect of tides on binary populations}",
-      journal = {Monthly Notices of the Royal Astronomical Society},
-     keywords = {METHODS: ANALYTICAL, METHODS: STATISTICAL, BINARIES: GENERAL, STARS: EVOLUTION, STARS: VARIABLES: OTHER, GALAXIES: STELLAR CONTENT, Astrophysics},
-         year = 2002,
-        month = feb,
-       volume = {329},
-       number = {4},
-        pages = {897-928},
-          doi = {10.1046/j.1365-8711.2002.05038.x},
-archivePrefix = {arXiv},
-       eprint = {astro-ph/0201220},
- primaryClass = {astro-ph},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2002MNRAS.329..897H},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@ARTICLE{Rodriguez:2021,
-       author = {{Rodriguez}, Carl L. and {Weatherford}, Newlin C. and {Coughlin}, Scott C. and {Amaro-Seoane}, Pau and {Breivik}, Katelyn and {Chatterjee}, Sourav and {Fragione}, Giacomo and {K{\i}ro{\u{g}}lu}, Fulya and {Kremer}, Kyle and {Rui}, Nicholas Z. and {Ye}, Claire S. and {Zevin}, Michael and {Rasio}, Frederic A.},
-        title = "{Modeling Dense Star Clusters in the Milky Way and beyond with the Cluster Monte Carlo Code}",
-      journal = {\apjs},
-     keywords = {1883, 1971, 678, 656, 1083, 1596, Astrophysics - Astrophysics of Galaxies, Astrophysics - Instrumentation and Methods for Astrophysics},
-         year = 2022,
-        month = feb,
-       volume = {258},
-       number = {2},
-          eid = {22},
-        pages = {22},
-          doi = {10.3847/1538-4365/ac2edf},
-archivePrefix = {arXiv},
-       eprint = {2106.02643},
- primaryClass = {astro-ph.GA},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2022ApJS..258...22R},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
-@book{Aarseth:2003, 
-      place={Cambridge}, 
-      series={Cambridge Monographs on Mathematical Physics}, 
[…]
