### 2025-12-04 eea30da
removed hanging print statement
