### 2025-12-04 3005ad0
added citation badge to readme
