+++ README.md
@@ -2,9 +2,16 @@
 
 **METISSE** is an acronym for **MET**hod of **I**nterpolation for **S**ingle **S**tar **E**volution. It is a code package designed for rapidly computing the evolution of a large number of stars by interpolating within a set of pre-computed evolutionary tracks. The input tracks for interpolation are computed from stellar structure and evolution codes such as MESA. It is important that the input tracks should be in EEP format, meaning that significant evolutionary points like zero-age main sequence (ZAMS) should occur at the same line number across each file. METISSE can be used as a standalone code or in conjunction with other codes for evolving populations of single and binary stars.
 
-The documentation for METISSE is available [here](https://metisse.readthedocs.io/en/latest/). 
+The documentation for METISSE is available at [](https://metisse.readthedocs.io/en/latest/). 
 
 
 Further details including code capabilties are described in: [Agrawal et al 2020](https://arxiv.org/abs/2005.13177) and [Agrawal et al 2023](https://arxiv.org/abs/2303.10187).
 
+
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.17650929.svg)](https://doi.org/10.5281/zenodo.17650929)
+
+
+<a href="https://www.tomwagg.com/software-citation-station/?auto-select=METISSE">
+    <img src="https://img.shields.io/badge/Cite-METISSE-5f65cd" />
+</a>
 
