### 2025-12-08 503a41b
Copy edits to JOSS paper
