### 2025-12-09 #8817 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
