### 2025-12-13 27ea3fc
Initial editorial changes for publication in JOSS
