### 2025-12-14 #8817 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
